refactor(views): split admin.py into a views/admin package
Mechanical move (Task 16, beautify-phase-0-1 plan): nx_lib/views/admin.py
(3480 lines) becomes nx_lib/views/admin/ with submodules overview.py,
organizations.py (incl. branding), clients.py, processes.py, system.py
(status + dev restart + maintenance), logs.py, users.py (sessions & users),
permissions.py (access control). Each submodule keeps its own
register_routes(app); the package __init__.py fans out to them in the
original registration order and re-exports every public name plus every
name the test suite monkeypatches on the package object. No Blueprints, no
endpoint renames, nx_lib/__init__.py unchanged.

Unlike Task 13 (generali), this split's test-edit strategy is retarget, not
avoid: the plan calls for pointing each patch at the submodule that now
owns the route, since a single shared nx_lib.views.admin.has_permission
binding can no longer stand in for eight independently-imported bindings.
Retargeted in tests/integration/test_admin_routes.py: the 5 named patch
strings (PATHS, engine_nexora_db, invalidate_branding,
invalidate_mapping_config, plus the setattr(admin_module, ...) forms of
engine_nexora_db/_client_codes) and all ~10 has_permission patches, each to
its owning submodule (system/clients/processes/users/organizations).
Attribute-of-object patches (PATHS.branding, engine_nexora_db.raw_connection,
clients_registry.CLIENTS/REGISTRY_DEGRADED_REASON) needed no retargeting --
same shared object regardless of import path.

Verified the retargeting has teeth: temporarily pointed
test_admin_add_user_duplicate_returns_409_or_500's has_permission patch at
the wrong submodule (system instead of users) and confirmed it now fails
with a real 403 (not a silent no-op), then reverted.

Also updates tests/unit/test_coverage_thresholds.py's views/admin.py entry
to per-submodule baselines, and fixes stale nx_lib/views/admin.py path
comments in nx_lib/clients.py, test_permission_guard.py and
test_create_app.py.

Full admin test file (189 tests) + tests/unit (1383 tests) green. Browser
smoke on /admin, /admin/organizations, /admin/clients, /admin/processes,
/admin/status, /admin/logs, /admin/sessions and /admin/access_control, no
console errors.
