--- nx_lib/views/admin.py
@@ -1,3479 +1,0 @@
-"""Admin pages: overview, organizations, maintenance banners, logs, sessions,
-user CRUD, access control, permissions."""
-
-import csv
-import math
-import os
-import re
-import secrets
-import subprocess
-from contextlib import suppress
-from datetime import datetime
-from pathlib import Path
-
-import bcrypt
-import pyodbc
-from flask import (
-    Response,
-    abort,
-    current_app,
-    jsonify,
-    redirect,
-    render_template,
-    request,
-    session,
-    url_for,
-)
-from flask_babel import gettext as _
-from werkzeug.exceptions import HTTPException
-from werkzeug.utils import secure_filename
-
-from .. import clients as clients_registry
-from .. import mapping_config, status
-from ..branding import _HEX_RE, invalidate_branding
-from ..branding import registry as branding_registry
-from ..clients import _ENGINE_KEYS
-from ..config import DOTENV_KEYS, IS_PROD, PATHS, REPO_ROOT
-from ..db import (
-    engine_generali_db,
-    engine_ms02_docfields_pg,
-    engine_ms02_pg,
-    engine_ms02_stats_pg,
-    engine_nexora_db,
-    engine_octo_db,
-    engine_statistics_db,
-    ping_dbs_parallel,
-)
-from ..files import is_file_allowed
-from ..maintenance import (
-    _MAINTENANCE_BLOCK_CACHE,
-    _get_blocking_maintenance,
-    _maintenance_parse_payload,
-    _maintenance_row_to_dict,
-)
-from ..mapping_config import invalidate_mapping_config
-from ..security import (
-    _revoke_session_by_id,
-    has_permission,
-    load_permissions_for_user,
-    page_visibility,
-    require_permission,
-)
-
-# ----------------------------------- overview ---------------------------------- #
-
-
-@require_permission("admin.view")
-def admin_dashboard():
-    if "username" not in session:
-        return redirect(url_for("login"))
-
-    user_count = 0
-    org_count = 0
-    active_sessions_count = None
-    failed_logins_today = None
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute("SELECT COUNT(*) FROM Users")
-        row = cursor.fetchone()
-        if row:
-            user_count = row[0]
-
-        cursor.execute("SELECT COUNT(*) FROM organizations")
-        row = cursor.fetchone()
-        if row:
-            org_count = row[0]
-
-        cursor.execute("""
-            SELECT COUNT(*) FROM ActiveSessions
-            WHERE LastSeenAt >= DATEADD(minute, -30, GETDATE())
-        """)
-        row = cursor.fetchone()
-        if row:
-            active_sessions_count = row[0]
-
-        cursor.execute("""
-            SELECT COUNT(*) FROM Logs
-            WHERE Path IN ('/login', '/verify_2fa')
-              AND HttpRequestMethod = 'POST'
-              AND HttpResponseCode >= 400
-              AND Timestamp >= CAST(GETDATE() AS DATE)
-        """)
-        row = cursor.fetchone()
-        if row:
-            failed_logins_today = row[0]
-    except Exception as e:
-        current_app.logger.error(f"Failed to load admin overview counts: {e}")
-    finally:
-        try:
-            if cursor:
-                cursor.close()
-        except Exception:
-            pass
-        try:
-            if conn:
-                conn.close()
-        except Exception:
-            pass
-
-    db_health = ping_dbs_parallel(
-        [
-            (engine_nexora_db, "Nexora"),
-            (engine_octo_db, "Octo"),
-            (engine_statistics_db, "Stats"),
-            (engine_generali_db, "Generali"),
-            *([(engine_ms02_pg, "MS02 (PG)")] if engine_ms02_pg is not None else []),
-            *(
-                [(engine_ms02_stats_pg, "MS02 stats (PG)")]
-                if engine_ms02_stats_pg is not None
-                else []
-            ),
-            *(
-                [(engine_ms02_docfields_pg, "MS02 docfields (PG)")]
-                if engine_ms02_docfields_pg is not None
-                else []
-            ),
-        ],
-        timeout_s=0.8,
-    )
-
-    return render_template(
-        "admin/admin_overview.html",
-        user_count=user_count,
-        org_count=org_count,
-        active_sessions_count=active_sessions_count,
-        failed_logins_today=failed_logins_today,
-        db_health=db_health,
-        current_env=os.environ.get("ENVIRONMENT", "?"),
-        can_restart=_restart_allowed(),
-        logged_in_user=session.get("username"),
-        userid=session.get("userid"),
-        page_visibility=page_visibility(),
-    )
-
-
-# ----------------------------------- organizations ---------------------------------- #
-
-
-@require_permission("admin.view.organizations")
-def admin_organizations_view():
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("select organizationcode, organization from organizations")
-        organizations = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-
-        # #98 phase 4: current branding per org, for the panel's initial state.
-        # registry() returns None when the load fails (or when the database
-        # predates migration 0081) -- degrade to "nothing branded", never error.
-        brands = branding_registry() or {}
-        for org in organizations:
-            brand = brands.get(org.get("organizationcode")) or {}
-            org["brand_name"] = brand.get("name")
-            org["brand_accent_hex"] = brand.get("accent_hex")
-            org["brand_logo_file"] = brand.get("logo_file")
-
-        return render_template(
-            "admin/organizations.html",
-            organizations=organizations,
-            can_edit_branding=has_permission("admin.edit.organization.branding"),
-            logged_in_user=session.get("username"),
-            userid=session.get("userid"),
-            page_visibility=page_visibility(),
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch organizations: {e}")
-        return render_template("500.html")
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.add.organization")
-def admin_add_organization():
-    import re as _re
-
-    data = request.get_json()
-    organization = data.get("organizationname")
-
-    if not organization:
-        return jsonify({"success": False, "message": _("All fields are required.")}), 400
-
-    clean_name = _re.sub(r"[^A-Z0-9]", "", organization.upper())
-    consonants = _re.sub(r"[AEIOU]", "", clean_name)
-    vowels = _re.sub(r"[^AEIOU]", "", clean_name)
-    code = consonants + vowels
-    organizationcode = code[:4].ljust(4, "X")
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("INSERT INTO organizations VALUES(?,?)", (organizationcode, organization))
-        conn.commit()
-        return jsonify({"success": True, "message": _("Organization created successfully.")})
-    except pyodbc.IntegrityError:
-        return jsonify({"success": False, "message": _("Organization already exists.")}), 409
-    except Exception as e:
-        current_app.logger.error(f"Error adding organization: {e}")
-        return jsonify({"success": False, "message": _("An unexpected error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.organization")
-def admin_edit_organization(organizationcode):
-    data = request.get_json()
-    organization = data.get("organizationname")
-    current_user_id = session["userid"]
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "UPDATE organizations SET organization=? WHERE organizationcode=?",
-            (organization, organizationcode),
-        )
-        conn.commit()
-        return jsonify({"success": True, "message": _("Organization updated successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error editing Organization {current_user_id}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.delete.organization")
-def admin_delete_organization(organizationcode):
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM organizations WHERE organizationcode=?", (organizationcode,))
-        conn.commit()
-
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Organization not found.")}), 404
-
-        # A delete mutates the branding registry exactly like a save does: skip
-        # this and the dead org keeps its brand -- and /branding/<code>/logo
-        # keeps serving its image -- for up to the 60s TTL. The file has to go
-        # too, or it is orphaned forever and would be re-exposed verbatim if the
-        # same org code is ever created again (#98 phase 4).
-        _delete_branding_logo(organizationcode)
-        invalidate_branding()
-
-        return jsonify({"success": True, "message": _("Organization deleted successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error deleting Organization {organizationcode}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.organizations")
-def api_admin_organizations_list():
-    if "username" not in session:
-        return jsonify({"error": "Not authorized"}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT organizationcode, organization FROM organizations ORDER BY organization"
-        )
-        orgs = [
-            dict(zip([c[0] for c in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify(orgs)
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch organizations list: {e}")
-        return jsonify({"error": _("An unexpected error occurred")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-# ------------------------------ organization branding ----------------------- #
-#
-# Branding attaches to the Organization (the customer -- PRVR, LKTR, ...),
-# never to ClientCode (the runtime source). See nx_lib/branding.py.
-
-BRANDING_MAX_BYTES = 512 * 1024
-# Narrows is_file_allowed's own table, which also knows pdf/xlsx. The ext check
-# short-circuits before the sniff, so a PDF never reaches the stream at all.
-BRANDING_LOGO_EXTS = ("svg", "png", "jpg", "jpeg")
-# The org code is used to build a filename, so it must be a plain code. Task 10's
-# serve route defends itself independently (os.path.basename on the stored name).
-_ORG_CODE_RE = re.compile(r"^[A-Za-z0-9]{1,16}$")
-
-
-def _branding_logo_target(branding_dir: Path, filename: str) -> Path | None:
-    """Resolve ``filename`` inside ``branding_dir``, or None if it escapes it.
-
-    The single path derivation + traversal check shared by the upload and the
-    delete side (the serve side in nx_lib/views/core.py defends itself with the
-    same basename/containment idiom against a hostile stored value)."""
-    target = (branding_dir / os.path.basename(filename)).resolve()
-    if target.parent != branding_dir.resolve():
-        return None
-    return target
-
-
-def _delete_branding_logo(organizationcode: str) -> None:
-    """Remove the logo files an organization's uploads left in var/branding/.
-
-    Sweeps every allowed extension rather than trusting BrandLogoFile: the row
-    is already gone by the time this runs, a NULL BrandLogoFile can still
-    coexist with a file on disk (switching formats leaves the old extension
-    behind), and every upload writes exactly ``<orgcode>.<ext>``. A missing
-    file is not an error."""
-    if not _ORG_CODE_RE.match(organizationcode or ""):
-        return
-    branding_dir = Path(PATHS.branding)
-    for ext in BRANDING_LOGO_EXTS:
-        target = _branding_logo_target(branding_dir, f"{organizationcode}.{ext}")
-        if target is None:
-            continue
-        try:
-            target.unlink()
-        except FileNotFoundError:
-            pass
-        except OSError as e:
-            current_app.logger.warning(f"branding logo unlink failed for {target}: {e}")
-
-
-def _org_exists(cursor, organizationcode):
-    cursor.execute(
-        "SELECT organizationcode FROM organizations WHERE organizationcode = ?",
-        (organizationcode,),
-    )
-    return cursor.fetchone() is not None
-
-
-@require_permission("admin.edit.organization.branding")
-def api_admin_organization_branding_save(organizationcode):
-    """Save an organization's brand name, accent hex and logo (#98 phase 4).
-
-    Accepts JSON (name/accent only) or multipart/form-data (plus ``logo``).
-
-    Upload safety is nx_lib/files.py's ``is_file_allowed`` -- secure_filename
-    plus a libmagic sniff of the actual bytes. The client-declared content type
-    is never consulted. SVG is allowed (spec D9) and is script-capable, which is
-    why branding_logo() serves every logo with a sandbox CSP + nosniff.
-
-    Path traversal: ``organizationcode`` reaches a filesystem path, so it is
-    matched against _ORG_CODE_RE first and the resolved target is re-checked to
-    live directly inside PATHS.branding. The stored filename is derived from the
-    org code, never from the uploaded filename.
-    """
-    if not _ORG_CODE_RE.match(organizationcode or ""):
-        return jsonify({"success": False, "message": _("Organization not found.")}), 404
-
-    data = (request.get_json(silent=True) or {}) if request.is_json else request.form
-
-    brand_name = (data.get("brand_name") or "").strip() or None
-    accent = (data.get("brand_accent_hex") or "").strip() or None
-    if accent is not None and not _HEX_RE.match(accent):
-        return (
-            jsonify(
-                {"success": False, "message": _("Accent colour must be a hex value like #336699.")}
-            ),
-            400,
-        )
-
-    logo = request.files.get("logo")
-    logo_bytes = None
-    logo_filename = None
-    if logo is not None and logo.filename:
-        safe_name = secure_filename(logo.filename)
-        ext = safe_name.rsplit(".", 1)[-1].lower() if "." in safe_name else ""
-        if ext not in BRANDING_LOGO_EXTS:
-            return (
-                jsonify(
-                    {"success": False, "message": _("Logo must be an SVG, PNG or JPEG image.")}
-                ),
-                400,
-            )
-        # Size before sniff, and both before anything touches the disk.
-        logo.stream.seek(0, os.SEEK_END)
-        size = logo.stream.tell()
-        logo.stream.seek(0)
-        if size > BRANDING_MAX_BYTES:
-            return (
-                jsonify({"success": False, "message": _("Logo is too large (max 512 KB).")}),
-                400,
-            )
-        if not is_file_allowed(safe_name, logo.stream):
-            return (
-                jsonify(
-                    {"success": False, "message": _("Logo must be an SVG, PNG or JPEG image.")}
-                ),
-                400,
-            )
-        logo_bytes = logo.stream.read()
-        logo_filename = f"{organizationcode}.{ext}"
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        if not _org_exists(cursor, organizationcode):
-            return jsonify({"success": False, "message": _("Organization not found.")}), 404
-
-        if logo_bytes is not None:
-            branding_dir = Path(PATHS.branding)
-            branding_dir.mkdir(parents=True, exist_ok=True)
-            target = _branding_logo_target(branding_dir, logo_filename)
-            if target is None:
-                return jsonify({"success": False, "message": _("Organization not found.")}), 404
-            target.write_bytes(logo_bytes)
-            cursor.execute(
-                "UPDATE organizations SET BrandName=?, BrandAccentHex=?, BrandLogoFile=? "
-                "WHERE organizationcode=?",
-                (brand_name, accent, logo_filename, organizationcode),
-            )
-        else:
-            cursor.execute(
-                "UPDATE organizations SET BrandName=?, BrandAccentHex=? WHERE organizationcode=?",
-                (brand_name, accent, organizationcode),
-            )
-        conn.commit()
-    except Exception as e:
-        current_app.logger.error(f"Error saving branding for {organizationcode}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-    # Must run on every successful save -- otherwise the edit looks broken for
-    # up to the registry's 60s TTL.
-    invalidate_branding()
-    return jsonify(
-        {
-            "success": True,
-            "message": _("Branding saved."),
-            # brand_logo_file is null when this save carried no upload: the
-            # stored logo is left untouched, so the caller keeps what it had.
-            "brand": {
-                "brand_name": brand_name,
-                "brand_accent_hex": accent,
-                "brand_logo_file": logo_filename,
-            },
-        }
-    )
-
-
-# ----------------------------------- clients (runtime sources) --------------------- #
-
-
-@require_permission("admin.view.clients")
-def admin_clients_view():
-    """List of dbo.Clients -- runtime sources (default/ms02), not customers
-    (see dbo.Organizations). The add/edit/delete affordances are rendered only
-    for ``admin.edit.clients`` (``can_edit``); the endpoints re-check it."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT ClientCode, DisplayName, Dialect, RuntimeEngineKey, StatsEngineKey, "
-            "StatsDialect, DocfieldsEngineKey, DocfieldsDialect, OctoDomain, SecretRef, "
-            "IsActive FROM dbo.Clients ORDER BY ClientCode"
-        )
-        clients = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        # Configured state (the table) is not resolved state (what the process
-        # actually runs on). 0079's seed is unconditional, so PROD gets an
-        # 'ms02' row whether or not env/PROD.env carries the MS02_* keys -- and
-        # without them _build_clients() skips it, leaving the page cheerfully
-        # reporting "Active: Yes" for a runtime that does not exist. Mark each
-        # row with whether the live registry actually holds it.
-        for client in clients:
-            client["loaded"] = client.get("ClientCode") in clients_registry.CLIENTS
-
-        return render_template(
-            "admin/clients.html",
-            clients=clients,
-            registry_degraded_reason=clients_registry.REGISTRY_DEGRADED_REASON,
-            can_edit=has_permission("admin.edit.clients"),
-            logged_in_user=session.get("username"),
-            userid=session.get("userid"),
-            page_visibility=page_visibility(),
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch clients: {e}")
-        return render_template("handlers/500.html"), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-_CLIENTS_ALLOWED_DIALECTS = {"tsql", "postgres"}
-_CLIENT_CODE_RE = re.compile(r"^[a-z0-9_]{2,50}$")
-# SecretRef holds only an env-key *prefix* (e.g. "MS02"), never a secret value
-# -- cfg.py looks up f"{SecretRef}_OCTO_CLIENT_SECRET" etc. This shape check is
-# the cheap guard called for: it rejects anything that isn't prefix-shaped
-# (lower-case, spaces, punctuation -- the kind of thing a pasted secret has),
-# not a guarantee that a value can never sneak in.
-_SECRET_REF_RE = re.compile(r"^[A-Z0-9_]{0,20}$")
-
-
-def _validate_client_payload(data, *, require_code):
-    """Server-side validation for the client add/edit endpoints -- never trust
-    the client-side checks in _clients_js.html. Returns a list of error
-    messages (empty when the payload is valid)."""
-    errors = []
-
-    client_code = (data.get("ClientCode") or "").strip()
-    if (require_code or client_code) and not _CLIENT_CODE_RE.match(client_code):
-        errors.append(_("Client code must be 2-50 lowercase letters, digits or underscores."))
-
-    if not (data.get("DisplayName") or "").strip():
-        errors.append(_("Display name is required."))
-
-    allowed = ", ".join(sorted(_CLIENTS_ALLOWED_DIALECTS))
-    dialect = data.get("Dialect")
-    if dialect not in _CLIENTS_ALLOWED_DIALECTS:
-        errors.append(_("Dialect must be one of: %(allowed)s", allowed=allowed))
-    for field in ("StatsDialect", "DocfieldsDialect"):
-        value = data.get(field)
-        if value and value not in _CLIENTS_ALLOWED_DIALECTS:
-            errors.append(_("%(field)s must be one of: %(allowed)s", field=field, allowed=allowed))
-
-    if data.get("RuntimeEngineKey") not in _ENGINE_KEYS:
-        errors.append(_("Unknown runtime engine key."))
-    for field in ("StatsEngineKey", "DocfieldsEngineKey"):
-        value = data.get(field)
-        if value and value not in _ENGINE_KEYS:
-            errors.append(_("Unknown %(field)s.", field=field))
-
-    secret_ref = data.get("SecretRef")
-    if secret_ref and not _SECRET_REF_RE.match(secret_ref):
-        errors.append(
-            _(
-                "Secret ref must be an env-key prefix (upper-case letters, digits, "
-                "underscores) -- never a secret value."
-            )
-        )
-
-    return errors
-
-
-@require_permission("admin.edit.clients")
-def api_admin_clients_add():
-    """Add a dbo.Clients row (migration 0079) -- a runtime source (default/
-    ms02: which DB/dialect/Octo tenant serves a client), not a customer
-    (dbo.Organizations). CLIENTS is built once at import (nx_lib/clients.py),
-    so a new row needs an app-pool recycle before it is picked up."""
-    data = request.get_json() or {}
-    errors = _validate_client_payload(data, require_code=True)
-    if errors:
-        return jsonify({"success": False, "message": " ".join(str(e) for e in errors)}), 400
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "INSERT INTO dbo.Clients (ClientCode, DisplayName, Dialect, RuntimeEngineKey, "
-            "StatsEngineKey, StatsDialect, DocfieldsEngineKey, DocfieldsDialect, OctoDomain, "
-            "SecretRef, IsActive) VALUES (?,?,?,?,?,?,?,?,?,?,?)",
-            (
-                data["ClientCode"].strip(),
-                data["DisplayName"].strip(),
-                data["Dialect"],
-                data["RuntimeEngineKey"],
-                data.get("StatsEngineKey") or None,
-                data.get("StatsDialect") or None,
-                data.get("DocfieldsEngineKey") or None,
-                data.get("DocfieldsDialect") or None,
-                data.get("OctoDomain") or None,
-                data.get("SecretRef") or None,
-                1 if data.get("IsActive", True) else 0,
-            ),
-        )
-        conn.commit()
-        return jsonify({"success": True, "message": _("Client created successfully.")})
-    except pyodbc.IntegrityError:
-        return jsonify({"success": False, "message": _("Client code already exists.")}), 409
-    except Exception as e:
-        current_app.logger.error(f"Error adding client: {e}")
-        return jsonify({"success": False, "message": _("An unexpected error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.clients")
-def api_admin_clients_edit(clientcode):
-    data = request.get_json() or {}
-    errors = _validate_client_payload(data, require_code=False)
-    if errors:
-        return jsonify({"success": False, "message": " ".join(str(e) for e in errors)}), 400
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "UPDATE dbo.Clients SET DisplayName=?, Dialect=?, RuntimeEngineKey=?, "
-            "StatsEngineKey=?, StatsDialect=?, DocfieldsEngineKey=?, DocfieldsDialect=?, "
-            "OctoDomain=?, SecretRef=?, IsActive=? WHERE ClientCode=?",
-            (
-                data["DisplayName"].strip(),
-                data["Dialect"],
-                data["RuntimeEngineKey"],
-                data.get("StatsEngineKey") or None,
-                data.get("StatsDialect") or None,
-                data.get("DocfieldsEngineKey") or None,
-                data.get("DocfieldsDialect") or None,
-                data.get("OctoDomain") or None,
-                data.get("SecretRef") or None,
-                1 if data.get("IsActive", True) else 0,
-                clientcode,
-            ),
-        )
-        conn.commit()
-
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Client not found.")}), 404
-
-        return jsonify({"success": True, "message": _("Client updated successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error editing client {clientcode}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.clients")
-def api_admin_clients_delete(clientcode):
-    """Refuses (409) a ClientCode still referenced by dbo.ProcessSources (migration
-    0074) instead of deleting it out from under the mapping-config registry
-    (nx_lib/mapping_config.py) or the running CLIENTS registry."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT COUNT(*) FROM dbo.ProcessSources WHERE ClientCode = ?", (clientcode,)
-        )
-        row = cursor.fetchone()
-        if row and row[0]:
-            return (
-                jsonify(
-                    {
-                        "success": False,
-                        "message": _(
-                            "This client is still referenced by process sources and "
-                            "cannot be deleted."
-                        ),
-                    }
-                ),
-                409,
-            )
-
-        cursor.execute("DELETE FROM dbo.Clients WHERE ClientCode=?", (clientcode,))
-        conn.commit()
-
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Client not found.")}), 404
-
-        return jsonify({"success": True, "message": _("Client deleted successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error deleting client {clientcode}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-# ----------------------------------- processes (mapping config) -------------------- #
-
-
-def _process_field_dict(field_mapping):
-    return {
-        "field_key": field_mapping.field_key,
-        "column": field_mapping.column,
-        "column_type": field_mapping.column_type,
-    }
-
-
-def _process_source_dict(source, fields):
-    return {
-        "client": source.client,
-        "process": source.process,
-        "table": source.table,
-        "alias": source.alias,
-        "join_condition": source.join_condition,
-        "time_filter": source.time_filter,
-        "suggestion_time_filter": source.suggestion_time_filter,
-        "export_column": source.export_column,
-        "import_column": source.import_column,
-        "workitem_column": source.workitem_column,
-        "extra_condition": source.extra_condition,
-        "id_column_type": source.id_column_type,
-        "fields": [_process_field_dict(m) for m in fields],
-    }
-
-
-@require_permission("admin.view.processes")
-def admin_processes_view():
-    """Read-only view of dbo.ProcessSources / ProcessFieldMappings (migration
-    0074), grouped by ClientCode -- a *runtime source* (default/ms02, see
-    dbo.Clients), not a customer (dbo.Organizations) -- then by ProcessName,
-    with each process's field mappings as a nested table. Read entirely
-    through the cached nx_lib/mapping_config.py registry, never raw SQL.
-
-    registry() returning None means the config failed to load (a load error
-    is never cached) -- render an explicit "unavailable" state rather than an
-    empty-looking success. The edit affordances render only for
-    admin.edit.processes; the free-form SQL fragment columns (JoinCondition,
-    TimeFilter, SuggestionTimeFilter, ExtraCondition) stay read-only for
-    everybody -- they are editable only by a migration."""
-    reg = mapping_config.registry()
-    clients_data = []
-    if reg is not None:
-        by_client = {}
-        for (client, process), source in reg.sources.items():
-            fields = sorted(
-                (m for m in reg.mappings if m.client == client and m.process == process),
-                key=lambda m: m.field_key,
-            )
-            by_client.setdefault(client, []).append((process, source, fields))
-        for client in sorted(by_client):
-            processes = sorted(by_client[client], key=lambda item: item[0])
-            clients_data.append(
-                {
-                    "client": client,
-                    "processes": [
-                        _process_source_dict(source, fields) for _, source, fields in processes
-                    ],
-                }
-            )
-
-    return render_template(
-        "admin/processes.html",
-        mapping_config_available=reg is not None,
-        can_edit=has_permission("admin.edit.processes"),
-        clients_data=clients_data,
-        client_codes=_client_codes(),
-        logged_in_user=session.get("username"),
-        userid=session.get("userid"),
-        page_visibility=page_visibility(),
-    )
-
-
-@require_permission("admin.view.processes")
-def api_admin_processes_list():
-    """JSON mirror of admin_processes_view() for a single client (or every
-    client when ``?client=`` is omitted) -- read through nx_lib/mapping_config.py,
-    never raw SQL. Mirrors that module's fail-closed contract: a registry load
-    failure is a 503, never an empty-looking 200 (task 6 brief)."""
-    client = request.args.get("client")
-    reg = mapping_config.registry()
-    if reg is None:
-        return jsonify({"error": _("Mapping config is currently unavailable.")}), 503
-
-    sources = mapping_config.sources_for(client)
-    processes = []
-    for source in sorted(sources, key=lambda s: (s.client, s.process)):
-        fields = sorted(
-            mapping_config.mappings_for(source.client, [source.process]),
-            key=lambda m: m.field_key,
-        )
-        processes.append(_process_source_dict(source, fields))
-
-    return jsonify({"client": client, "processes": processes})
-
-
-# --------------------- processes (mapping config, writes) -------------------------- #
-
-# ProcessName MUST be exactly <customer>.<process> -- two dot-separated
-# segments, no more, no fewer. This is NOT cosmetic and NOT "convention only":
-# every consumer of the auto-provisioned workitems.filter.process.<ProcessName>
-# permission reconstructs the process name from the permission code as exactly
-# the LAST TWO dot-segments (nx_lib/views/workitems.py, nx_lib/process_helpers.py
-# `parts[-2], parts[-1]`). A one-segment name ("Invoice") derives back as
-# "process.Invoice" and the grant silently never matches; a three-segment name
-# ("acme.eu.01_Invoice") derives back as "eu.01_Invoice", same silent dead end;
-# and worse, "x.acme.01_Invoice" reduces to "acme.01_Invoice", so it would
-# piggyback on another customer's existing grant. Until this page existed the
-# invariant held only because process names were migration-controlled -- now an
-# admin types them, so it is enforced here.
-_PROCESS_NAME_RE = re.compile(r"^[A-Za-z0-9_\-]{1,49}\.[A-Za-z0-9_\-]{1,50}$")
-
-# FieldKey is a plain mapping key (doctype, invoice_no) -- it is never turned
-# into a permission code, so it keeps the looser shape.
-_FIELD_KEY_RE = re.compile(r"^[A-Za-z0-9_.\-]{1,100}$")
-
-# TableName / TableAlias / ColumnName / Export|Import|WorkitemColumn are
-# INTERPOLATED into SQL by the downstream query builders (f-strings in
-# nx_lib/views/dashboard.py and nx_lib/workitem_sources.py) -- they are config,
-# not bind parameters. Every admin-editable field that reaches a builder is
-# validated on the way in, here, server-side. Quotes and brackets are allowed
-# because the live config holds qualified identifiers in both dialects
-# (dbo.tblAlpha, public."DossierStatistik") -- spaces, semicolons, parentheses
-# and comment markers are not.
-_IDENT = re.compile(r'^[A-Za-z_][A-Za-z0-9_."\[\]]{0,99}$')
-
-# ColumnType / IdColumnType are never interpolated -- they are only compared
-# against the literal type buckets in workitems.py / workitem_sources.py (e.g.
-# 'nvarchar', 'character varying'), hence the space.
-_COLUMN_TYPE_RE = re.compile(r"^[A-Za-z][A-Za-z0-9_ ]{0,29}$")
-
-# ProcessSources.JoinCondition / TimeFilter / SuggestionTimeFilter /
-# ExtraCondition are free-form SQL fragments by design and are deliberately NOT
-# writable here: no validator can make an arbitrary predicate safe, so they
-# stay migration-only. They appear in no INSERT/UPDATE and in no form below --
-# a payload carrying them is ignored, not applied.
-_PROCESS_PERMISSION_PREFIX = "workitems.filter.process."
-
-
-def _validate_identifier_fields(data, fields, errors, *, required=()):
-    """Append an error for every ``fields`` entry that is present but is not a
-    plain SQL identifier (and for every ``required`` entry that is missing)."""
-    for field in fields:
-        value = (data.get(field) or "").strip()
-        if not value:
-            if field in required:
-                errors.append(_("%(field)s is required.", field=field))
-            continue
-        if not _IDENT.match(value):
-            errors.append(
-                _(
-                    "%(field)s must be a plain SQL identifier (letters, digits, "
-                    "underscores, dots, quotes or brackets).",
-                    field=field,
-                )
-            )
-
-
-def _validate_process_identity(client_code, process_name, errors, field_key=None):
-    if not _CLIENT_CODE_RE.match(client_code or ""):
-        errors.append(_("Client code must be 2-50 lowercase letters, digits or underscores."))
-    if not _PROCESS_NAME_RE.match(process_name or ""):
-        errors.append(
-            _(
-                "Process name must be exactly <customer>.<process> -- two parts separated "
-                "by a single dot (e.g. acme.01_Invoice), letters, digits, dashes or "
-                "underscores only. The permission that grants access to this process is "
-                "derived from those two parts, so any other shape can never be granted."
-            )
-        )
-    if field_key is not None and not _FIELD_KEY_RE.match(field_key or ""):
-        errors.append(_("Field key must be 1-100 letters, digits, dots, dashes or underscores."))
-
-
-def _validate_process_source_payload(data, client_code, process_name):
-    """Server-side validation for the process-source write endpoints -- never
-    trust the client-side checks in templates/js/admin/_processes_js.html."""
-    errors = []
-    _validate_process_identity(client_code, process_name, errors)
-    _validate_identifier_fields(
-        data,
-        ("TableName", "TableAlias", "ExportColumn", "ImportColumn", "WorkitemColumn"),
-        errors,
-        required=("TableName",),
-    )
-    if len((data.get("TableAlias") or "").strip()) > 10:
-        errors.append(_("Table alias must be at most 10 characters."))
-    id_column_type = (data.get("IdColumnType") or "").strip()
-    if id_column_type and not _COLUMN_TYPE_RE.match(id_column_type):
-        errors.append(_("Id column type must be a plain SQL type name."))
-    return errors
-
-
-def _validate_field_mapping_payload(data, client_code, process_name, field_key):
-    errors = []
-    _validate_process_identity(client_code, process_name, errors, field_key=field_key)
-    _validate_identifier_fields(data, ("ColumnName",), errors, required=("ColumnName",))
-    column_type = (data.get("ColumnType") or "").strip()
-    if column_type and not _COLUMN_TYPE_RE.match(column_type):
-        errors.append(_("Column type must be a plain SQL type name."))
-    return errors
-
-
-def _process_source_values(data):
-    """The six writable ProcessSources columns, in INSERT/UPDATE order. The
-    free-form SQL fragment columns are absent on purpose."""
-    return (
-        (data.get("TableName") or "").strip(),
-        (data.get("TableAlias") or "").strip() or None,
-        (data.get("ExportColumn") or "").strip() or None,
-        (data.get("ImportColumn") or "").strip() or None,
-        (data.get("WorkitemColumn") or "").strip() or None,
-        (data.get("IdColumnType") or "").strip() or None,
-    )
-
-
-def _validation_error(errors):
-    return jsonify({"success": False, "message": " ".join(str(e) for e in errors)}), 400
-
-
-def _permission_reduction(process_name):
-    """The (customer, process) pair every consumer derives back out of a
-    ``workitems.filter.process.<ProcessName>`` code -- the last two dot
-    segments. Two process names sharing a reduction share an entitlement,
-    whatever their ClientCode: the permission code carries no client."""
-    return ".".join((process_name or "").split(".")[-2:])
-
-
-def _reduction_conflict(cursor, client_code, process_name):
-    """The existing (ClientCode, ProcessName) whose permission reduction
-    collides with ``process_name``, or None.
-
-    _PROCESS_NAME_RE already forces new names to two segments, so the new name
-    IS its own reduction -- but rows written before this page existed (or by a
-    migration) may have more, and a legacy 'x.acme.01_Invoice' reduces onto a
-    freshly typed 'acme.01_Invoice'. The exact same pair is not a conflict: the
-    PK insert below turns that into the usual 409 "already exists".
-    """
-    reduction = _permission_reduction(process_name)
-    cursor.execute("SELECT ClientCode, ProcessName FROM dbo.ProcessSources")
-    for row in cursor.fetchall():
-        existing = (row[0], row[1])
-        if existing == (client_code, process_name):
-            continue
-        if _permission_reduction(row[1]) == reduction:
-            return existing
-    return None
-
-
-def _client_code_exists(cursor, client_code):
-    cursor.execute("SELECT COUNT(*) FROM dbo.Clients WHERE ClientCode = ?", (client_code,))
-    row = cursor.fetchone()
-    return bool(row and row[0])
-
-
-def _client_codes():
-    """ClientCodes from dbo.Clients, for the /admin/processes picker. Read from
-    the table rather than from the resolved CLIENTS registry on purpose: a row
-    that is configured but not loaded (missing env keys) is still a legitimate
-    target for process config. Degrades to [] -- the picker then falls back to
-    free text rather than blocking the page."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT ClientCode FROM dbo.Clients ORDER BY ClientCode")
-        return [row[0] for row in cursor.fetchall()]
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch client codes: {e}")
-        return []
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_process_source_add():
-    """Add a dbo.ProcessSources row (migration 0074) AND provision its
-    ``workitems.filter.process.<ProcessName>`` permission in the same
-    transaction -- self-service onboarding is the whole point of this page, and
-    a process nobody can be granted is a half-created process.
-
-    The permission is created granted to NOBODY: granting stays a deliberate
-    act at /admin/access-control. The insert mirrors migration 0059's
-    idempotent WHERE NOT EXISTS shape, so re-adding a process whose permission
-    row outlived an earlier delete is a no-op rather than a unique-key error."""
-    data = request.get_json() or {}
-    client_code = (data.get("ClientCode") or "").strip()
-    process_name = (data.get("ProcessName") or "").strip()
-    errors = _validate_process_source_payload(data, client_code, process_name)
-    if errors:
-        return _validation_error(errors)
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        # Nothing in the schema ties ProcessSources.ClientCode to dbo.Clients
-        # (migration 0079 deliberately adds no FK -- ProcessSources predates the
-        # table), so a typo like 'defualt' would otherwise create a process
-        # source, auto-provision its permission, and yield config that can never
-        # resolve. Check it here instead.
-        if not _client_code_exists(cursor, client_code):
-            return _validation_error(
-                [
-                    _(
-                        "Unknown client code %(code)s -- it must be an existing "
-                        "runtime source from the Clients page.",
-                        code=client_code,
-                    )
-                ]
-            )
-        conflict = _reduction_conflict(cursor, client_code, process_name)
-        if conflict:
-            return (
-                jsonify(
-                    {
-                        "success": False,
-                        "message": _(
-                            "Process name %(name)s collides with the existing "
-                            "%(other_client)s / %(other)s: both resolve to the same "
-                            "permission %(code)s, so a grant for one would entitle the "
-                            "other. Pick a different name.",
-                            name=process_name,
-                            other_client=conflict[0],
-                            other=conflict[1],
-                            code=f"{_PROCESS_PERMISSION_PREFIX}{_permission_reduction(process_name)}",
-                        ),
-                    }
-                ),
-                409,
-            )
-        cursor.execute(
-            "INSERT INTO dbo.ProcessSources (ClientCode, ProcessName, TableName, TableAlias, "
-            "ExportColumn, ImportColumn, WorkitemColumn, IdColumnType) VALUES (?,?,?,?,?,?,?,?)",
-            (client_code, process_name, *_process_source_values(data)),
-        )
-        code = f"{_PROCESS_PERMISSION_PREFIX}{process_name}"
-        cursor.execute(
-            "INSERT INTO dbo.Permission (Code, Description) SELECT ?, ? "
-            "WHERE NOT EXISTS (SELECT 1 FROM dbo.Permission p WHERE p.Code = ?)",
-            (code, f"View {process_name} workitems"[:200], code),
-        )
-        conn.commit()
-        invalidate_mapping_config()
-        return jsonify(
-            {
-                "success": True,
-                "message": _(
-                    "Process source created. Its permission %(code)s was created but "
-                    "granted to nobody -- grant it under Access Control.",
-                    code=code,
-                ),
-            }
-        )
-    except pyodbc.IntegrityError:
-        return (
-            jsonify({"success": False, "message": _("This process source already exists.")}),
-            409,
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error adding process source {client_code}/{process_name}: {e}")
-        return jsonify({"success": False, "message": _("An unexpected error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_process_source_edit(clientcode, processname):
-    """Edit the identifier columns of a dbo.ProcessSources row. Identity
-    (ClientCode, ProcessName) comes from the URL and is never rewritten -- a
-    rename would strand the process's field mappings and its permission."""
-    data = request.get_json() or {}
-    errors = _validate_process_source_payload(data, clientcode, processname)
-    if errors:
-        return _validation_error(errors)
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "UPDATE dbo.ProcessSources SET TableName=?, TableAlias=?, ExportColumn=?, "
-            "ImportColumn=?, WorkitemColumn=?, IdColumnType=? "
-            "WHERE ClientCode=? AND ProcessName=?",
-            (*_process_source_values(data), clientcode, processname),
-        )
-        conn.commit()
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Process source not found.")}), 404
-        invalidate_mapping_config()
-        return jsonify({"success": True, "message": _("Process source updated successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error editing process source {clientcode}/{processname}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_process_source_delete(clientcode, processname):
-    """Delete a dbo.ProcessSources row. Refused with 409 while field mappings
-    still reference it -- FK_ProcessFieldMappings_ProcessSources would raise
-    anyway, and a 500 tells the admin nothing about what to do next.
-
-    The ``workitems.filter.process.<name>`` permission row is deliberately left
-    behind: dropping it would silently revoke access the admin never asked to
-    change, and re-adding the process re-uses it (see the add endpoint)."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT COUNT(*) FROM dbo.ProcessFieldMappings WHERE ClientCode = ? "
-            "AND ProcessName = ?",
-            (clientcode, processname),
-        )
-        row = cursor.fetchone()
-        if row and row[0]:
-            return (
-                jsonify(
-                    {
-                        "success": False,
-                        "message": _(
-                            "This process source still has field mappings and cannot be "
-                            "deleted. Remove them first."
-                        ),
-                    }
-                ),
-                409,
-            )
-
-        cursor.execute(
-            "DELETE FROM dbo.ProcessSources WHERE ClientCode=? AND ProcessName=?",
-            (clientcode, processname),
-        )
-        conn.commit()
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Process source not found.")}), 404
-        invalidate_mapping_config()
-        return jsonify({"success": True, "message": _("Process source deleted successfully.")})
-    except pyodbc.IntegrityError:
-        return (
-            jsonify(
-                {
-                    "success": False,
-                    "message": _("This process source is still referenced and cannot be deleted."),
-                }
-            ),
-            409,
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error deleting process source {clientcode}/{processname}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_field_mapping_add():
-    """Add a dbo.ProcessFieldMappings row -- one doc-field of one process."""
-    data = request.get_json() or {}
-    client_code = (data.get("ClientCode") or "").strip()
-    process_name = (data.get("ProcessName") or "").strip()
-    field_key = (data.get("FieldKey") or "").strip()
-    errors = _validate_field_mapping_payload(data, client_code, process_name, field_key)
-    if errors:
-        return _validation_error(errors)
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "INSERT INTO dbo.ProcessFieldMappings (ClientCode, ProcessName, FieldKey, "
-            "ColumnName, ColumnType) VALUES (?,?,?,?,?)",
-            (
-                client_code,
-                process_name,
-                field_key,
-                (data.get("ColumnName") or "").strip(),
-                (data.get("ColumnType") or "").strip() or None,
-            ),
-        )
-        conn.commit()
-        invalidate_mapping_config()
-        return jsonify({"success": True, "message": _("Field mapping created successfully.")})
-    except pyodbc.IntegrityError:
-        # PK_ProcessFieldMappings, or FK_ProcessFieldMappings_ProcessSources
-        # when the parent process source does not exist.
-        return (
-            jsonify(
-                {
-                    "success": False,
-                    "message": _(
-                        "This field mapping already exists, or its process source does not."
-                    ),
-                }
-            ),
-            409,
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error adding field mapping {client_code}/{process_name}: {e}")
-        return jsonify({"success": False, "message": _("An unexpected error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_field_mapping_edit(clientcode, processname, fieldkey):
-    """Edit a field mapping's column. Identity comes from the URL -- renaming a
-    FieldKey is a delete plus an add, not an update."""
-    data = request.get_json() or {}
-    errors = _validate_field_mapping_payload(data, clientcode, processname, fieldkey)
-    if errors:
-        return _validation_error(errors)
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "UPDATE dbo.ProcessFieldMappings SET ColumnName=?, ColumnType=? "
-            "WHERE ClientCode=? AND ProcessName=? AND FieldKey=?",
-            (
-                (data.get("ColumnName") or "").strip(),
-                (data.get("ColumnType") or "").strip() or None,
-                clientcode,
-                processname,
-                fieldkey,
-            ),
-        )
-        conn.commit()
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Field mapping not found.")}), 404
-        invalidate_mapping_config()
-        return jsonify({"success": True, "message": _("Field mapping updated successfully.")})
-    except Exception as e:
-        current_app.logger.error(
-            f"Error editing field mapping {clientcode}/{processname}/{fieldkey}: {e}"
-        )
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.processes")
-def api_admin_field_mapping_delete(clientcode, processname, fieldkey):
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "DELETE FROM dbo.ProcessFieldMappings WHERE ClientCode=? AND ProcessName=? "
-            "AND FieldKey=?",
-            (clientcode, processname, fieldkey),
-        )
-        conn.commit()
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Field mapping not found.")}), 404
-        invalidate_mapping_config()
-        return jsonify({"success": True, "message": _("Field mapping deleted successfully.")})
-    except Exception as e:
-        current_app.logger.error(
-            f"Error deleting field mapping {clientcode}/{processname}/{fieldkey}: {e}"
-        )
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-# ----------------------------------- status page ---------------------------------- #
-
-
-@require_permission("admin.status.view")
-def admin_status_view():
-    """Component health + incident history, as recorded by ops/outage_monitor.py.
-
-    Deliberately reads only what the monitor persisted (issue #167) instead of
-    probing on page load: the point of the page is "what has been true since
-    yesterday evening", which a request-scoped ping cannot answer. When the
-    monitor has not reported recently the page says so rather than rendering a
-    reassuring all-green grid from stale rows.
-    """
-    try:
-        data = status.load_status(
-            engine_nexora_db,
-            maintenance=_get_blocking_maintenance() is not None,
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to load status page data: {e}")
-        return render_template("handlers/500.html"), 500
-    return render_template(
-        "admin/status.html",
-        status=data,
-        # Import-time degradation of the client registry (nx_lib/clients.py):
-        # its only other signal is a logger.error that fires before Flask has
-        # configured logging, so it never reaches app.log.
-        clients_registry_degraded_reason=clients_registry.REGISTRY_DEGRADED_REASON,
-        stale_after_min=status.DEFAULT_STALE_AFTER_S // 60,
-        logged_in_user=session.get("username"),
-        userid=session.get("userid"),
-        page_visibility=page_visibility(),
-    )
-
-
-# ----------------------------------- dev server restart ---------------------------------- #
-
-
-_SWITCHABLE_ENVS = {"INT", "STAGING"}
-
-
-def _restart_allowed():
-    """May this caller restart / env-switch the dev server?
-
-    The ``admin.restart`` permission lives in NexoraDB, and STAGING resolves
-    NexoraDB to the prod server (DB_SERVER_PRD) where the row from migration
-    0059 doesn't exist — so on STAGING the control vanished and the env switch
-    was one-way (#198). Loopback callers are therefore allowed regardless of the
-    permission, the same trust rule the /dev/* routes use (#193). PROD is out
-    either way: it's IIS-hosted, where restarting means recycling the app pool.
-    """
-    if IS_PROD:
-        return False
-    return has_permission("admin.restart") or request.remote_addr in ("127.0.0.1", "::1")
-
-
-def api_admin_restart():
-    """Restart the local dev server process (issue #184). Dev-only — 404s on PROD
-    since PROD is IIS-hosted and restarting there means recycling the app pool,
-    not killing a `python nx_main.py` process. Fires bin/nx.ps1 -r as a hidden
-    background process; it kills this process and starts a fresh one, so the response has
-    to make it back to the browser before that happens (nx.ps1 sleeps ~1s first).
-
-    Optional JSON body {"env": "INT"|"STAGING"} switches ENVIRONMENT on the way
-    back up (issue #187) via nx.ps1's existing --env: flag — PROD is refused by
-    nx.ps1 itself, so no need to re-check it here."""
-    if IS_PROD:
-        abort(404)
-    if not _restart_allowed():
-        abort(403)
-    target_env = (request.get_json(silent=True) or {}).get("env")
-    if target_env is not None and target_env not in _SWITCHABLE_ENVS:
-        return jsonify({"success": False, "message": _("Unknown environment.")}), 400
-    args = ["pwsh", "-File", str(REPO_ROOT / "bin" / "nx.ps1"), "-r"]
-    if target_env:
-        args.append(f"--env:{target_env}")
-    # restart the instance we're actually serving from — without this a
-    # --no-conflict instance's restart button would kill the port-8000 one
-    args.append(f"--port:{request.environ.get('SERVER_PORT', '8000')}")
-    subprocess.Popen(
-        args,
-        cwd=str(REPO_ROOT),
-        # Strip the keys load_dotenv injected into this process: inherited,
-        # they'd win over the target env file in the new server (override=False)
-        # and it would run INT DB connections while claiming to be STAGING.
-        env={k: v for k, v in os.environ.items() if k not in DOTENV_KEYS},
-        # CREATE_NO_WINDOW, not DETACHED_PROCESS: pwsh exits 0 without running
-        # the script when it has no console at all (#187); a hidden console
-        # works and still survives this process being killed by nx.ps1.
-        creationflags=subprocess.CREATE_NO_WINDOW | subprocess.CREATE_NEW_PROCESS_GROUP,
-    )
-    return jsonify({"success": True})
-
-
-# ----------------------------------- maintenance banner ---------------------------------- #
-
-
-@require_permission("admin.maintenance.view")
-def admin_maintenance_view():
-    try:
-        return render_template(
-            "admin/maintenance.html",
-            logged_in_user=session.get("username"),
-            userid=session.get("userid"),
-            page_visibility=page_visibility(),
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to load maintenance page: {e}")
-        return render_template("handlers/500.html"), 500
-
-
-@require_permission("admin.maintenance.view")
-def api_admin_maintenance_list():
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("""
-            SELECT ID, Title, Message, StartAt, EndAt, Severity, Active, BlockAccess, AnnounceMinutesBefore, CreatedBy, CreatedAt
-            FROM MaintenanceBanner
-            ORDER BY StartAt DESC, ID DESC
-        """)
-        cols = [c[0] for c in cursor.description]
-        records = [_maintenance_row_to_dict(row, cols) for row in cursor.fetchall()]
-        return jsonify({"success": True, "records": records})
-    except Exception as e:
-        current_app.logger.error(f"Maintenance list error: {e}")
-        return jsonify({"success": False, "error": _("An unexpected error occurred")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.maintenance.edit")
-def api_admin_maintenance_add():
-    body = request.get_json(force=True) or {}
-    parsed, err = _maintenance_parse_payload(body)
-    if err:
-        return jsonify({"success": False, "error": err[0]}), err[1]
-
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            INSERT INTO MaintenanceBanner (Title, Message, StartAt, EndAt, Severity, Active, BlockAccess, AnnounceMinutesBefore, CreatedBy, CreatedAt)
-            OUTPUT INSERTED.ID
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, GETDATE())
-        """,
-            [
-                parsed["title"],
-                parsed["message"],
-                parsed["start_at"],
-                parsed["end_at"],
-                parsed["severity"],
-                parsed["active"],
-                parsed["block_access"],
-                parsed["announce_minutes"],
-                session.get("userid"),
-            ],
-        )
-        new_id = cursor.fetchone()[0]
-        conn.commit()
-        _MAINTENANCE_BLOCK_CACHE["expires_at"] = 0.0
-        return jsonify({"success": True, "id": int(new_id)})
-    except Exception as e:
-        current_app.logger.error(f"Maintenance add error: {e}")
-        return jsonify({"success": False, "error": _("An unexpected error occurred")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.maintenance.edit")
-def api_admin_maintenance_edit(banner_id):
-    body = request.get_json(force=True) or {}
-    parsed, err = _maintenance_parse_payload(body)
-    if err:
-        return jsonify({"success": False, "error": err[0]}), err[1]
-
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            UPDATE MaintenanceBanner
-               SET Title = ?, Message = ?, StartAt = ?, EndAt = ?, Severity = ?,
-                   Active = ?, BlockAccess = ?, AnnounceMinutesBefore = ?
-             WHERE ID = ?
-        """,
-            [
-                parsed["title"],
-                parsed["message"],
-                parsed["start_at"],
-                parsed["end_at"],
-                parsed["severity"],
-                parsed["active"],
-                parsed["block_access"],
-                parsed["announce_minutes"],
-                banner_id,
-            ],
-        )
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "error": "Banner not found"}), 404
-        conn.commit()
-        _MAINTENANCE_BLOCK_CACHE["expires_at"] = 0.0
-        return jsonify({"success": True})
-    except Exception as e:
-        current_app.logger.error(f"Maintenance edit error: {e}")
-        return jsonify({"success": False, "error": _("An unexpected error occurred")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.maintenance.edit")
-def api_admin_maintenance_delete(banner_id):
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM MaintenanceBanner WHERE ID = ?", [banner_id])
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "error": "Banner not found"}), 404
-        conn.commit()
-        _MAINTENANCE_BLOCK_CACHE["expires_at"] = 0.0
-        return jsonify({"success": True})
-    except Exception as e:
-        current_app.logger.error(f"Maintenance delete error: {e}")
-        return jsonify({"success": False, "error": _("An unexpected error occurred")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-# ----------------------------------- logs ---------------------------------- #
-
-
-@require_permission("admin.view.system.logs")
-def admin_logs_view():
-    organizations = []
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("select organizationcode, organization from organizations")
-        organizations = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-    except Exception as e:
-        current_app.logger.error(f"Failed to load organizations for logs page: {e}")
-    finally:
-        try:
-            if cursor:
-                cursor.close()
-        except Exception:
-            pass
-        try:
-            if conn:
-                conn.close()
-        except Exception:
-            pass
-    return render_template(
-        "admin/logs.html",
-        organizations=organizations,
-        logged_in_user=session.get("username"),
-        userid=session.get("userid"),
-        page_visibility=page_visibility(),
-    )
-
-
-def _build_logs_where_clause():
-    """Pull filter args from request and produce (where_clause, params).
-    Shared between /api/admin/logs/search and /api/admin/logs/export.csv."""
-    username = request.args.get("username", "").strip()
-    method = request.args.get("method", "").strip()
-    path = request.args.get("path", "").strip()
-    status = request.args.get("status", "").strip()
-    start_date = request.args.get("start_date", "")
-    end_date = request.args.get("end_date", "")
-    organization = request.args.get("organization", "").strip()
-
-    parts = ["1=1"]
-    params = []
-
-    if username:
-        parts.append("Username LIKE ?")
-        params.append(f"%{username}%")
-    if method:
-        parts.append("HttpRequestMethod = ?")
-        params.append(method)
-    if path:
-        parts.append("Path LIKE ?")
-        params.append(f"%{path}%")
-    if status == "SUCCESS":
-        parts.append("HttpResponseCode BETWEEN 200 AND 299")
-    elif status == "FAILURE":
-        parts.append("HttpResponseCode >= 400")
-    if start_date:
-        parts.append("Timestamp >= ?")
-        params.append(start_date)
-    if end_date:
-        parts.append("Timestamp <= ?")
-        # Day-granularity filters (e.g. "Today", "Last 7 days") send a bare
-        # "YYYY-MM-DD" date and rely on us rounding up to end-of-day. Sub-day
-        # presets (e.g. "Last hour") send a full "YYYY-MM-DD HH:MM:SS"
-        # timestamp already — don't append another time onto it.
-        end_bound = end_date if " " in end_date else f"{end_date} 23:59:59"
-        params.append(end_bound)
-    if organization:
-        parts.append("Username IN (SELECT username FROM Users WHERE organizationcode = ?)")
-        params.append(organization)
-
-    return " AND ".join(parts), params
-
-
-@require_permission("admin.view.system.logs")
-def api_admin_logs_search():
-    page = request.args.get("page", 1, type=int)
-    per_page = 50
-    offset = (page - 1) * per_page
-
-    where_clause, params = _build_logs_where_clause()
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute(f"SELECT COUNT(*) FROM Logs WHERE {where_clause}", params)
-        total_count = cursor.fetchone()[0]
-
-        sql = f"""
-            SELECT LogID, Timestamp, Username, HttpRequestMethod, Path,
-                   HttpResponseCode, Args, RequestIpAddress, durationSeconds
-            FROM Logs
-            WHERE {where_clause}
-            ORDER BY Timestamp DESC
-            OFFSET ? ROWS FETCH NEXT ? ROWS ONLY
-        """
-        cursor.execute(sql, [*params, offset, per_page])
-
-        logs = []
-        for row in cursor.fetchall():
-            logs.append(
-                {
-                    "LogID": row.LogID,
-                    # Emit ISO-8601 explicitly so the client can pass it straight
-                    # to `new Date(...)`. Flask's default JSON encoder uses RFC 1123
-                    # which doesn't survive the +'Z' timezone-suffix hack.
-                    "Timestamp": row.Timestamp.isoformat()
-                    if hasattr(row.Timestamp, "isoformat")
-                    else row.Timestamp,
-                    "Username": row.Username,
-                    "HttpRequestMethod": row.HttpRequestMethod,
-                    "Path": row.Path,
-                    "HttpResponseCode": row.HttpResponseCode,
-                    "Args": row.Args,
-                    "RequestIpAddress": row.RequestIpAddress,
-                    "durationSeconds": row.durationSeconds,
-                }
-            )
-
-        return jsonify(
-            {
-                "logs": logs,
-                "total": total_count,
-                "page": page,
-                "pages": math.ceil(total_count / per_page),
-            }
-        )
-    except Exception as e:
-        current_app.logger.error(f"Log search error: {e}")
-        return jsonify({"error": _("An unexpected error occurred")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.system.logs")
-def api_admin_logs_export():
-    """Stream the filtered log set as CSV. Capped at 50k rows so a wide-open
-    filter doesn't yank the whole table."""
-    max_rows = 50000
-    where_clause, params = _build_logs_where_clause()
-
-    sql = f"""
-        SELECT TOP ({max_rows}) LogID, Timestamp, Username, HttpRequestMethod, Path,
-               HttpResponseCode, RequestIpAddress, durationSeconds, Args
-        FROM Logs
-        WHERE {where_clause}
-        ORDER BY Timestamp DESC
-    """
-
-    def generate():
-        import io as _io
-
-        buf = _io.StringIO()
-        writer = csv.writer(buf, quoting=csv.QUOTE_MINIMAL)
-        writer.writerow(
-            [
-                "LogID",
-                "Timestamp",
-                "Username",
-                "Method",
-                "Path",
-                "StatusCode",
-                "IPAddress",
-                "DurationSeconds",
-                "Args",
-            ]
-        )
-        yield buf.getvalue()
-        buf.seek(0)
-        buf.truncate(0)
-
-        conn = engine_nexora_db.raw_connection()
-        try:
-            cursor = conn.cursor()
-            cursor.execute(sql, params)
-            while True:
-                rows = cursor.fetchmany(500)
-                if not rows:
-                    break
-                for row in rows:
-                    ts = row.Timestamp
-                    if ts is None:
-                        ts_iso = ""
-                    elif hasattr(ts, "isoformat"):
-                        ts_iso = ts.isoformat()
-                    else:
-                        ts_iso = str(ts).replace(" ", "T", 1)
-                    writer.writerow(
-                        [
-                            row.LogID,
-                            ts_iso,
-                            row.Username or "",
-                            row.HttpRequestMethod or "",
-                            row.Path or "",
-                            row.HttpResponseCode if row.HttpResponseCode is not None else "",
-                            row.RequestIpAddress or "",
-                            row.durationSeconds if row.durationSeconds is not None else "",
-                            row.Args or "",
-                        ]
-                    )
-                yield buf.getvalue()
-                buf.seek(0)
-                buf.truncate(0)
-            cursor.close()
-        finally:
-            conn.close()
-
-    filename = f"nexora-logs-{datetime.now().strftime('%Y%m%d-%H%M%S')}.csv"
-    return Response(
-        generate(),
-        mimetype="text/csv; charset=utf-8",
-        headers={
-            "Content-Disposition": f'attachment; filename="{filename}"',
-            "Cache-Control": "no-store",
-        },
-    )
-
-
-# ----------------------------------- sessions & users ---------------------------------- #
-
-
-@require_permission("admin.view.active.sessions")
-def admin_sessions_view():
-    return render_template(
-        "admin/sessions.html",
-        logged_in_user=session.get("username"),
-        userid=session.get("userid"),
-        page_visibility=page_visibility(),
-    )
-
-
-@require_permission("admin.create.user")
-def admin_add_user():
-    from .auth import _build_reset_email_message, send_reset_email
-
-    data = request.get_json()
-    username = data.get("username")
-    password = data.get("password")
-    fullname = data.get("fullname")
-    email = data.get("email")
-    organization = data.get("organization")
-    accessprofile = data.get("accessprofile")
-    # Checkbox: an unticked box is simply absent from the posted form.
-    send_invite = bool(data.get("send_invite"))
-
-    if send_invite:
-        # Nobody -- not the admin, not the mail -- ever sees this one. It is a
-        # placeholder that keeps the account unusable until the invited user
-        # sets their own password through the emailed link.
-        password = secrets.token_urlsafe(32)
-
-    if not all([username, password, fullname, email, organization, accessprofile]):
-        return jsonify({"success": False, "message": _("All fields are required.")}), 400
-
-    if not has_permission(f"admin.assign.user.accessprofile.{str(accessprofile).lower()}"):
-        current_app.logger.error(
-            "assign-permission denied: profile=%r username=%r",
-            str(accessprofile)[:100],
-            str(username)[:100],
-        )
-        return jsonify({"success": False, "message": _("Permission Denied for this action.")}), 403
-
-    hashed_password = bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt()).decode("utf-8")
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("select accessid from accessprofile where name = ?", accessprofile)
-        accessid = cursor.fetchone()[0]
-        cursor.execute(
-            "select organizationcode from organizations where organization = ?", organization
-        )
-        organizationcode = cursor.fetchone()[0]
-        cursor.execute(
-            "INSERT INTO Users (username, password, fullname, email, organizationcode, accessid, InitReset) VALUES (?, ?, ?, ?, ?, ?, ?)",
-            (
-                username,
-                hashed_password,
-                fullname,
-                email,
-                organizationcode,
-                accessid,
-                # An invited user picks their own password through the emailed
-                # link, so the forced first-login change (InitReset NULL) would
-                # only make them do it twice. They still land on 2FA enrolment.
-                1 if send_invite else None,
-            ),
-        )
-        conn.commit()
-
-        if send_invite:
-            # Synchronous on purpose: the admin needs to know whether the mail
-            # actually went out. No timing oracle to dodge here (unlike the
-            # self-service reset), and Graph is capped at 10s per call.
-            try:
-                sent = send_reset_email(
-                    email, message=_build_reset_email_message(email, invite=True)
-                )
-            except Exception as e:
-                current_app.logger.error(f"Failed to send invite email to {email}: {e}")
-                sent = False
-            if not sent:
-                return jsonify(
-                    {
-                        "success": True,
-                        "message": _(
-                            "User created, but the email could not be sent. "
-                            "Ask them to use 'Forgot password'."
-                        ),
-                    }
-                )
-            return jsonify({"success": True, "message": _("User created and email sent.")})
-
-        return jsonify({"success": True, "message": _("User created successfully.")})
-    except pyodbc.IntegrityError:
-        return jsonify({"success": False, "message": _("Username or email already exists.")}), 409
-    except Exception as e:
-        current_app.logger.error(f"Error adding user: {e}")
-        return jsonify({"success": False, "message": _("An unexpected error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.user")
-def admin_edit_user(user_id):
-    data = request.get_json()
-    username = data.get("username")
-    fullname = data.get("fullname")
-    email = data.get("email")
-    password = data.get("password")
-    organization = data.get("organization")
-    accessprofile = data.get("accessprofile")
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        # The <select> only ever lists this admin's own assignable profiles
-        # (see admin_user_detail). If the edited user's CURRENT profile isn't
-        # in that set, the template still renders it as a selected-but-
-        # unassignable option so an untouched form round-trips the same
-        # value. Only require the assign-permission when the value actually
-        # CHANGES — leaving it alone must never 403 or silently reassign it.
-        cursor.execute(
-            "select ap.name from users u left join accessprofile ap on u.accessid = ap.accessid where u.userid = ?",
-            user_id,
-        )
-        row = cursor.fetchone()
-        current_profile = row[0] if row else None
-
-        # Defense-in-depth: the template always submits SOME accessprofile
-        # value now (the current one, if the admin never touched the
-        # dropdown -- see user_detail.html). A client that still omits the
-        # key entirely (JSON body with no "accessprofile", not the normal
-        # browser path) must be treated as "leave it unchanged", not as
-        # "clear the profile" -- the latter would 403 an admin trying to
-        # save an unrelated field.
-        if accessprofile is None:
-            accessprofile = current_profile
-
-        if accessprofile != current_profile and not has_permission(
-            f"admin.assign.user.accessprofile.{str(accessprofile).lower()}"
-        ):
-            current_app.logger.error(
-                f"User does not have Permission: admin.assign.user.accessprofile.{str(accessprofile).lower()} for {user_id}"
-            )
-            return jsonify(
-                {"success": False, "message": _("Permission Denied for this action.")}
-            ), 403
-
-        cursor.execute("select accessid from accessprofile where name = ?", accessprofile)
-        accessid = cursor.fetchone()[0]
-        cursor.execute(
-            "select organizationcode from organizations where organization = ?", organization
-        )
-        organizationcode = cursor.fetchone()[0]
-
-        if password:
-            hashed_password = bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt()).decode(
-                "utf-8"
-            )
-            cursor.execute(
-                "UPDATE Users SET username=?, fullname=?, email=?, password=?, organizationcode=?,accessid=? WHERE userID=?",
-                (username, fullname, email, hashed_password, organizationcode, accessid, user_id),
-            )
-        else:
-            cursor.execute(
-                "UPDATE Users SET username=?, fullname=?, email=?,  organizationcode=?,accessid=? WHERE userID=?",
-                (username, fullname, email, organizationcode, accessid, user_id),
-            )
-        conn.commit()
-
-        return jsonify({"success": True, "message": _("User updated successfully.")})
-    except Exception as e:
-        current_app.logger.error(f"Error editing user {user_id}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def admin_user_detail(user_id):
-    if "username" not in session:
-        return redirect(url_for("login"))
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-            SELECT u.userID, u.username, u.fullname, u.email,
-                   ap.Name AS AccessProfileName,
-                   o.organization, o.organizationcode
-            FROM Users u
-            LEFT JOIN AccessProfile ap ON u.accessid = ap.AccessID
-            LEFT JOIN Organizations o ON u.organizationcode = o.organizationcode
-            WHERE u.userID = ?
-        """,
-            (user_id,),
-        )
-        row = cursor.fetchone()
-        if not row:
-            abort(404)
-        user = dict(zip([c[0] for c in cursor.description], row, strict=False))
-
-        cursor.execute(
-            "SELECT organizationcode, organization FROM Organizations ORDER BY organization"
-        )
-        organizations = [
-            dict(zip([c[0] for c in cursor.description], r, strict=False))
-            for r in cursor.fetchall()
-        ]
-
-        cursor.execute(
-            "SELECT ap.name profile, ap.accessid accessid FROM AccessProfile ap ORDER BY ap.name"
-        )
-        all_ap = [
-            dict(zip([c[0] for c in cursor.description], r, strict=False))
-            for r in cursor.fetchall()
-        ]
-        assignable_profiles = [
-            ap
-            for ap in all_ap
-            if has_permission(f'admin.assign.user.accessprofile.{str(ap["profile"]).lower()}')
-        ]
-
-        cursor.execute("""
-            SELECT PermissionID, Code, Description FROM Permission
-            ORDER BY
-                LEFT(Code, LEN(Code) - CHARINDEX('.', REVERSE(Code))),
-                CASE
-                    WHEN Code LIKE '%.view'    THEN 1
-                    WHEN Code LIKE '%.add'     THEN 2
-                    WHEN Code LIKE '%.add.%'   THEN 3
-                    WHEN Code LIKE '%.edit%'   THEN 4
-                    WHEN Code LIKE '%.delete%' THEN 5
-                    ELSE 6
-                END,
-                Code
-        """)
-        all_permissions = [
-            dict(zip([c[0] for c in cursor.description], r, strict=False))
-            for r in cursor.fetchall()
-        ]
-
-        return render_template(
-            "admin/user_detail.html",
-            user=user,
-            organizations=organizations,
-            assignable_profiles=assignable_profiles,
-            all_permissions=all_permissions,
-            can_edit_user=has_permission("admin.edit.user"),
-            can_delete_user=has_permission("admin.delete.user"),
-            can_edit_overrides=has_permission("admin.edit.user.override"),
-            logged_in_user=session.get("username"),
-            userid=session.get("userid"),
-            page_visibility=page_visibility(),
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        current_app.logger.error(f"Error loading user detail {user_id}: {e}")
-        return render_template("500.html")
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_user_activity(user_id):
-    """Recent log entries for one user. Last 7 days, paginated, 25 per page."""
-    page = request.args.get("page", 1, type=int)
-    per_page = 25
-    offset = max(0, (page - 1) * per_page)
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute("SELECT username FROM Users WHERE userID = ?", (user_id,))
-        row = cursor.fetchone()
-        if not row:
-            return jsonify({"entries": [], "total": 0, "page": page, "pages": 0})
-        username = row[0]
-
-        cursor.execute(
-            """
-            SELECT COUNT(*) FROM Logs
-            WHERE Username = ?
-              AND Timestamp >= DATEADD(day, -7, GETDATE())
-        """,
-            (username,),
-        )
-        total = (cursor.fetchone() or [0])[0] or 0
-
-        cursor.execute(
-            """
-            SELECT Timestamp, HttpRequestMethod, Path, HttpResponseCode
-            FROM Logs
-            WHERE Username = ?
-              AND Timestamp >= DATEADD(day, -7, GETDATE())
-            ORDER BY Timestamp DESC
-            OFFSET ? ROWS FETCH NEXT ? ROWS ONLY
-        """,
-            (username, offset, per_page),
-        )
-
-        entries = []
-        for r in cursor.fetchall():
-            ts = r.Timestamp
-            if ts is None:
-                ts_iso = None
-            elif hasattr(ts, "isoformat"):
-                ts_iso = ts.isoformat()
-            else:
-                ts_iso = str(ts).replace(" ", "T", 1)
-            entries.append(
-                {
-                    "Timestamp": ts_iso,
-                    "HttpRequestMethod": r.HttpRequestMethod,
-                    "Path": r.Path,
-                    "HttpResponseCode": r.HttpResponseCode,
-                }
-            )
-
-        pages = max(1, math.ceil(total / per_page)) if total else 0
-
-        return jsonify(
-            {
-                "entries": entries,
-                "total": int(total),
-                "page": page,
-                "pages": pages,
-            }
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to load activity for user {user_id}: {e}")
-        return jsonify(
-            {
-                "error": _("An unexpected error occurred"),
-                "entries": [],
-                "total": 0,
-                "page": page,
-                "pages": 0,
-            }
-        ), 500
-    finally:
-        if cursor:
-            with suppress(Exception):
-                cursor.close()
-        if conn:
-            with suppress(Exception):
-                conn.close()
-
-
-@require_permission("admin.delete.user")
-def admin_delete_user(user_id):
-    current_user = session.get("userid")
-    if str(user_id) == current_user:
-        return jsonify({"success": False, "message": _("You cannot delete your own account.")}), 403
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        # Every child-row delete and the final Users delete run in ONE
-        # transaction committed exactly once at the end. NexoraDB's FKs to
-        # dbo.Users are all NO ACTION (no DB-level cascade), so the child rows
-        # must be removed first. The previous code committed each child delete
-        # individually, so any later failure — including the Users delete
-        # itself — left a half-deleted, undeletable user. Ordering only
-        # requires that every child delete precede `delete from users`.
-        cursor.execute("delete from userpermissionoverride where userid = ?", (user_id,))
-
-        # Reporting artifacts. FK_Reports_Users / FK_ReportSchedules_Users /
-        # FK_ReportShares_Users are all NO ACTION and were previously omitted,
-        # so deleting a report-owning user failed outright on the Users delete.
-        # Remove, in FK-safe order: the shares the user holds and the schedules
-        # they own; then any shares/schedules that reference reports the user
-        # OWNS (regardless of who holds/owns them); then the reports; then the
-        # Users row. Deleting the shares and schedules explicitly keeps this
-        # correct even though FK_ReportShares_Reports / FK_ReportSchedules_
-        # Reports are ON DELETE CASCADE — it does not rely on the DB cascade.
-        cursor.execute("delete from ReportShares where SharedWithUserID = ?", (user_id,))
-        cursor.execute(
-            "delete from ReportShares where ReportID in "
-            "(select ReportID from Reports where OwnerUserID = ?)",
-            (user_id,),
-        )
-        cursor.execute("delete from ReportSchedules where OwnerUserID = ?", (user_id,))
-        cursor.execute(
-            "delete from ReportSchedules where ReportID in "
-            "(select ReportID from Reports where OwnerUserID = ?)",
-            (user_id,),
-        )
-        cursor.execute("delete from Reports where OwnerUserID = ?", (user_id,))
-
-        cursor.execute("delete from users where userid = ?", (user_id,))
-        deleted = cursor.rowcount
-
-        conn.commit()
-
-        if deleted == 0:
-            return jsonify({"success": False, "message": _("User not found.")}), 404
-
-        return jsonify({"success": True, "message": _("User deleted successfully.")})
-    except Exception as e:
-        if conn is not None:
-            with suppress(Exception):
-                conn.rollback()
-        current_app.logger.error(f"Error deleting user {user_id}: {e}")
-        return jsonify({"success": False, "message": _("An error occurred.")}), 500
-    finally:
-        if cursor:
-            with suppress(Exception):
-                cursor.close()
-        if conn:
-            with suppress(Exception):
-                conn.close()
-
-
-@require_permission("admin.edit.user.override")
-def admin_revoke_session(session_id):
-    try:
-        _revoke_session_by_id(session_id)
-        return jsonify({"success": True})
-    except Exception as e:
-        current_app.logger.error(f"Failed to revoke session {session_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-
-
-@require_permission("admin.edit.user.override")
-def admin_revoke_all_sessions(user_id):
-    sids = []
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT SessionID FROM ActiveSessions WHERE UserID = ?", (user_id,))
-        sids = [row[0] for row in cursor.fetchall()]
-    except Exception as e:
-        current_app.logger.error(f"Failed to list sessions for user {user_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            with suppress(Exception):
-                cursor.close()
-        if conn:
-            with suppress(Exception):
-                conn.close()
-
-    revoked = 0
-    for sid in sids:
-        try:
-            if _revoke_session_by_id(sid):
-                revoked += 1
-        except Exception as e:
-            current_app.logger.warning(f"Revoke failed for session {sid}: {e}")
-
-    current_app.logger.info(
-        f"Admin {session.get('username')} revoked {revoked} session(s) for user {user_id}"
-    )
-    return jsonify({"success": True, "revoked": revoked})
-
-
-@require_permission("admin.view.users")
-def api_admin_users_list():
-    if "username" not in session:
-        return jsonify({"error": "Not authorized"}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("""
-            SELECT userID, username, fullname, email, ap.name accessprofile, o.organization organization
-            FROM Users u
-            JOIN accessprofile ap ON ap.accessid = u.accessid
-            JOIN organizations o ON o.organizationcode = u.organizationcode
-            ORDER BY username
-        """)
-        users = [
-            dict(zip([c[0] for c in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify(users)
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch users list: {e}")
-        return jsonify({"error": _("An unexpected error occurred")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.active.sessions")
-def admin_recent_logs():
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("""
-            SELECT TOP 20 Timestamp, Username, HttpRequestMethod, Path, HttpResponseCode
-            FROM Logs
-            ORDER BY Timestamp DESC
-        """)
-        logs = []
-        for row in cursor.fetchall():
-            # HttpResponseCode is NVARCHAR in the DB; coerce defensively rather
-            # than compare a string against int bounds (guaranteed TypeError).
-            try:
-                status_code = int(row.HttpResponseCode)
-            except (TypeError, ValueError):
-                status_code = None
-            logs.append(
-                {
-                    # Emit ISO-8601 explicitly so the client can pass it straight
-                    # to `new Date(...)`. Flask's default JSON encoder uses RFC 1123
-                    # which doesn't survive the +'Z' timezone-suffix hack.
-                    "Timestamp": row.Timestamp.isoformat()
-                    if hasattr(row.Timestamp, "isoformat")
-                    else row.Timestamp,
-                    "Username": row.Username,
-                    "ActionType": f"{row.HttpRequestMethod} {row.Path}",
-                    "ActionStatus": "SUCCESS"
-                    if status_code is not None and 200 <= status_code < 300
-                    else "FAILURE",
-                }
-            )
-        return jsonify(logs)
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch recent logs: {e}")
-        return jsonify({"error": _("Could not fetch logs")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.active.sessions")
-def admin_active_sessions():
-    """Read currently-active sessions from ActiveSessions, joined to Users.
-    Filtered to LastSeenAt (bumped on every request by _enforce_active_session)
-    so this reflects actual recent activity, not just login time (issue #109)."""
-    conn = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("""
-            SELECT
-                a.SessionID,
-                a.UserID    AS Userid,
-                u.username  AS Username,
-                a.IPAddress,
-                a.CreatedAt AS LoggedInAt
-            FROM ActiveSessions a
-            LEFT JOIN Users u ON u.userID = a.UserID
-            WHERE a.LastSeenAt >= DATEADD(minute, -30, GETDATE())
-            ORDER BY a.CreatedAt DESC
-        """)
-        sessions = []
-        for r in cursor.fetchall():
-            sessions.append(
-                {
-                    "SessionID": r[0],
-                    "Userid": r[1],
-                    "Username": r[2],
-                    "IPAddress": r[3],
-                    # Emit ISO-8601 explicitly so the client can pass it straight
-                    # to `new Date(...)`. Flask's default JSON encoder uses RFC 1123
-                    # which doesn't survive the +'Z' timezone-suffix hack.
-                    "LoggedInAt": r[4].isoformat() if r[4] else None,
-                }
-            )
-        return jsonify(sessions)
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch active sessions: {e}")
-        return jsonify({"error": _("Could not fetch sessions")}), 500
-    finally:
-        if conn:
-            conn.close()
-
-
-# ----------------------------------- access control & permissions ---------------------------------- #
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def admin_access_control():
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute("""
-            SELECT ap.AccessID, ap.Name, ap.Description, COUNT(u.userID) AS UserCount
-            FROM AccessProfile ap
-            LEFT JOIN Users u ON u.accessID = ap.AccessID
-            GROUP BY ap.AccessID, ap.Name, ap.Description
-            ORDER BY ap.Name
-        """)
-        profiles = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-
-        cursor.execute("""
-            SELECT PermissionID, Code, Description FROM Permission
-            ORDER BY
-                LEFT(Code, LEN(Code) - CHARINDEX('.', REVERSE(Code))),
-                CASE
-                    WHEN Code LIKE '%.view'    THEN 1
-                    WHEN Code LIKE '%.add'     THEN 2
-                    WHEN Code LIKE '%.add.%'   THEN 3
-                    WHEN Code LIKE '%.edit%'   THEN 4
-                    WHEN Code LIKE '%.delete%' THEN 5
-                    ELSE 6
-                END,
-                Code
-        """)
-        all_permissions = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-
-        organizations = []
-        assignable_profiles = []
-        if has_permission("admin.view.users"):
-            cursor.execute(
-                "SELECT organizationcode, organization FROM Organizations ORDER BY organization"
-            )
-            organizations = [
-                dict(zip([column[0] for column in cursor.description], row, strict=False))
-                for row in cursor.fetchall()
-            ]
-            cursor.execute(
-                "SELECT ap.name profile, ap.accessid accessid FROM AccessProfile ap ORDER BY ap.name"
-            )
-            all_ap = [
-                dict(zip([column[0] for column in cursor.description], row, strict=False))
-                for row in cursor.fetchall()
-            ]
-            for ap in all_ap:
-                if has_permission(f'admin.assign.user.accessprofile.{str(ap["profile"]).lower()}'):
-                    assignable_profiles.append(ap)
-
-        return render_template(
-            "admin/access_control.html",
-            profiles=profiles,
-            all_permissions=all_permissions,
-            organizations=organizations,
-            assignable_profiles=assignable_profiles,
-            can_edit_accessprofile=has_permission("admin.edit.accessprofile"),
-            can_view_users=has_permission("admin.view.users"),
-            can_create_user=has_permission("admin.create.user"),
-            can_edit_user=has_permission("admin.edit.user"),
-            can_delete_user=has_permission("admin.delete.user"),
-            logged_in_user=session.get("username"),
-            userid=session.get("userid"),
-            page_visibility=page_visibility(),
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error loading access control: {e}")
-        return render_template("500.html")
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def admin_permission_matrix():
-    """Read-only user x permission matrix (#172) -- filterable either
-    direction. Editing stays on access_control / user_detail; this page only
-    links through to them."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute("""
-            SELECT PermissionID, Code, Description FROM Permission
-            ORDER BY
-                LEFT(Code, LEN(Code) - CHARINDEX('.', REVERSE(Code))),
-                CASE
-                    WHEN Code LIKE '%.view'    THEN 1
-                    WHEN Code LIKE '%.add'     THEN 2
-                    WHEN Code LIKE '%.add.%'   THEN 3
-                    WHEN Code LIKE '%.edit%'   THEN 4
-                    WHEN Code LIKE '%.delete%' THEN 5
-                    ELSE 6
-                END,
-                Code
-        """)
-        all_permissions = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-
-        cursor.execute("""
-            SELECT u.userID, u.username, u.Fullname, u.organizationCode, o.organization
-            FROM Users u
-            LEFT JOIN Organizations o ON o.organizationcode = u.organizationCode
-            ORDER BY u.Fullname
-        """)
-        all_users = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-
-        return render_template(
-            "admin/permission_matrix.html",
-            all_permissions=all_permissions,
-            all_users=all_users,
-            page_visibility=page_visibility(),
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error loading permission matrix: {e}")
-        return render_template("500.html")
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_permission_holders(permission_id):
-    """Mirror of api_admin_user_effective_permissions with the axes flipped:
-    one permission, resolved across every user."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute(
-            "SELECT PermissionID, Code, Description FROM Permission WHERE PermissionID = ?",
-            (permission_id,),
-        )
-        p = cursor.fetchone()
-        if not p:
-            return jsonify({"success": False, "message": "Permission not found"}), 404
-
-        cursor.execute(
-            """
-            SELECT u.userID, u.username, u.Fullname, u.organizationCode, o.organization,
-                   ap_perm.Effect AS ProfileEffect,
-                   uo.Effect      AS OverrideEffect
-            FROM Users u
-            LEFT JOIN Organizations o ON o.organizationcode = u.organizationCode
-            LEFT JOIN AccessProfilePermission ap_perm
-                   ON ap_perm.AccessID = u.accessid
-                  AND ap_perm.PermissionID = ?
-            LEFT JOIN UserPermissionOverride uo
-                   ON uo.UserID = u.userID
-                  AND uo.PermissionID = ?
-            ORDER BY u.Fullname
-        """,
-            (permission_id, permission_id),
-        )
-
-        holders = []
-        for r in cursor.fetchall():
-            override = r.OverrideEffect
-            profile = r.ProfileEffect
-            if override == "A":
-                source = "override-allow"
-            elif override == "D":
-                continue  # explicitly denied -- not a holder
-            elif profile == "A":
-                source = "profile"
-            else:
-                continue  # no grant, no override -- not a holder
-
-            holders.append(
-                {
-                    "userID": r.userID,
-                    "username": r.username,
-                    "fullname": r.Fullname,
-                    "organizationCode": r.organizationCode,
-                    "organization": r.organization,
-                    "source": source,
-                }
-            )
-
-        return jsonify(
-            {
-                "success": True,
-                "permission": {
-                    "PermissionID": p.PermissionID,
-                    "Code": p.Code,
-                    "Description": p.Description,
-                },
-                "holders": holders,
-            }
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to compute holders for permission {permission_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            with suppress(Exception):
-                cursor.close()
-        if conn:
-            with suppress(Exception):
-                conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def get_users_admin_access_control():
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-
-    profile_filter = request.args.get("profile", "").strip()
-    org_filter = request.args.get("organization", "").strip()
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        where_parts = ["1=1"]
-        params = []
-        if profile_filter:
-            where_parts.append("ap.Name = ?")
-            params.append(profile_filter)
-        if org_filter:
-            where_parts.append("o.organizationcode = ?")
-            params.append(org_filter)
-        where_clause = " AND ".join(where_parts)
-
-        query = f"""
-            SELECT
-                u.userID, u.username, u.fullname, u.email,
-                ap.Name AS AccessProfileName, ap.AccessID AS AccessProfileID,
-                o.organizationcode,
-                o.organization,
-                (SELECT COUNT(*) FROM UserPermissionOverride upo WHERE upo.UserID = u.userID) AS OverrideCount
-            FROM Users u
-            LEFT JOIN AccessProfile ap ON u.accessid = ap.AccessID
-            LEFT JOIN Organizations o ON u.organizationcode = o.organizationcode
-            WHERE {where_clause}
-            ORDER BY u.fullname
-        """
-        cursor.execute(query, params)
-
-        users = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify(users)
-    except Exception as e:
-        current_app.logger.error(f"Failed to fetch users for access control: {e}")
-        return jsonify({"error": _("Could not fetch users")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def get_profile_details(access_id):
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT PermissionID, Effect
-            FROM AccessProfilePermission
-            WHERE AccessID = ?
-            """,
-            (access_id,),
-        )
-        assigned_perms = [
-            dict(zip([column[0] for column in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify({"success": True, "permissions": assigned_perms})
-    except Exception as e:
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.accessprofile")
-def save_access_profile():
-    data = request.get_json()
-    access_id = data.get("accessId")
-    name = data.get("name")
-    description = data.get("description")
-    permissions = data.get("permissions")
-
-    if not name:
-        return jsonify({"success": False, "message": _("Name is required")}), 400
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        if access_id:
-            cursor.execute(
-                "UPDATE AccessProfile SET Name=?, Description=? WHERE AccessID=?",
-                (name, description, access_id),
-            )
-            cursor.execute("DELETE FROM AccessProfilePermission WHERE AccessID=?", (access_id,))
-        else:
-            cursor.execute(
-                "INSERT INTO AccessProfile (Name, Description) OUTPUT INSERTED.AccessID VALUES (?, ?)",
-                (name, description),
-            )
-            access_id = cursor.fetchone()[0]
-
-        if permissions:
-            params = [(access_id, p["PermissionID"], p["Effect"]) for p in permissions]
-            cursor.executemany(
-                "INSERT INTO AccessProfilePermission (AccessID, PermissionID, Effect) VALUES (?, ?, ?)",
-                params,
-            )
-        conn.commit()
-        session["permissions"] = load_permissions_for_user(session["userid"])
-        return jsonify({"success": True, "message": _("Profile saved successfully")})
-    except Exception as e:
-        current_app.logger.error(f"Error saving profile: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def get_user_overrides(user_id):
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT AccessID FROM Users WHERE UserID = ?", (user_id,))
-        row = cursor.fetchone()
-        if not row:
-            return jsonify({"success": False, "message": "User not found"}), 404
-        base_access_id = row[0]
-        cursor.execute(
-            "SELECT PermissionID, Effect FROM UserPermissionOverride WHERE UserID = ?", (user_id,)
-        )
-        overrides = {row.PermissionID: row.Effect for row in cursor.fetchall()}
-        base_perms = {}
-        if base_access_id:
-            cursor.execute(
-                "SELECT PermissionID, Effect FROM AccessProfilePermission WHERE AccessID = ?",
-                (base_access_id,),
-            )
-            base_perms = {row.PermissionID: row.Effect for row in cursor.fetchall()}
-        return jsonify(
-            {
-                "success": True,
-                "overrides": overrides,
-                "base_permissions": base_perms,
-            }
-        )
-    except Exception as e:
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_user_effective_permissions(user_id):
-    """Compute the merged permission set: profile-grant unless an override
-    flips it. Source on each entry tells the UI whether it came from the
-    profile or from an Allow override."""
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-            SELECT u.userID, u.username, u.AccessID, ap.Name AS ProfileName
-            FROM Users u
-            LEFT JOIN AccessProfile ap ON ap.AccessID = u.AccessID
-            WHERE u.userID = ?
-        """,
-            (user_id,),
-        )
-        u = cursor.fetchone()
-        if not u:
-            return jsonify({"success": False, "message": "User not found"}), 404
-
-        cursor.execute(
-            """
-            SELECT p.PermissionID, p.Code, p.Description,
-                   ap_perm.Effect AS ProfileEffect,
-                   uo.Effect      AS OverrideEffect
-            FROM Permission p
-            LEFT JOIN AccessProfilePermission ap_perm
-                   ON ap_perm.PermissionID = p.PermissionID
-                  AND ap_perm.AccessID = ?
-            LEFT JOIN UserPermissionOverride uo
-                   ON uo.PermissionID = p.PermissionID
-                  AND uo.UserID = ?
-            ORDER BY
-                LEFT(p.Code, LEN(p.Code) - CHARINDEX('.', REVERSE(p.Code))),
-                CASE
-                    WHEN p.Code LIKE '%.view'    THEN 1
-                    WHEN p.Code LIKE '%.add'     THEN 2
-                    WHEN p.Code LIKE '%.add.%'   THEN 3
-                    WHEN p.Code LIKE '%.edit%'   THEN 4
-                    WHEN p.Code LIKE '%.delete%' THEN 5
-                    ELSE 6
-                END,
-                p.Code
-        """,
-            (u.AccessID, user_id),
-        )
-
-        granted = []
-        denied = []
-        for r in cursor.fetchall():
-            override = r.OverrideEffect
-            profile = r.ProfileEffect
-            if override == "A":
-                source = "override-allow"
-                effective = True
-            elif override == "D":
-                source = "override-deny"
-                effective = False
-            elif profile == "A":
-                source = "profile"
-                effective = True
-            elif profile == "D":
-                source = "profile-deny"
-                effective = False
-            else:
-                continue  # No grant, no override — irrelevant
-
-            entry = {
-                "PermissionID": r.PermissionID,
-                "Code": r.Code,
-                "Description": r.Description,
-                "source": source,
-            }
-            (granted if effective else denied).append(entry)
-
-        return jsonify(
-            {
-                "success": True,
-                "user": {
-                    "userID": u.userID,
-                    "username": u.username,
-                    "profile": u.ProfileName,
-                },
-                "granted": granted,
-                "denied": denied,
-            }
-        )
-    except Exception as e:
-        current_app.logger.error(f"Failed to compute effective permissions for user {user_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            with suppress(Exception):
-                cursor.close()
-        if conn:
-            with suppress(Exception):
-                conn.close()
-
-
-@require_permission("admin.edit.user.override")
-def save_user_overrides():
-    data = request.get_json()
-    user_id = data.get("userId")
-    overrides = data.get("overrides")
-
-    if not user_id:
-        return jsonify({"success": False, "message": "User ID required"}), 400
-
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-
-        cursor.execute("DELETE FROM UserPermissionOverride WHERE UserID=?", (user_id,))
-
-        if overrides:
-            params = [(user_id, p["PermissionID"], p["Effect"]) for p in overrides]
-            cursor.executemany(
-                "INSERT INTO UserPermissionOverride (UserID, PermissionID, Effect) VALUES (?, ?, ?)",
-                params,
-            )
-
-        conn.commit()
-        session["permissions"] = load_permissions_for_user(session["userid"])
-        return jsonify({"success": True, "message": _("Overrides updated successfully")})
-    except Exception as e:
-        current_app.logger.error(f"Error saving overrides: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_permissions_list():
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("""
-            SELECT
-                p.PermissionID, p.Code, p.Description,
-                (SELECT COUNT(*) FROM AccessProfilePermission a WHERE a.PermissionID = p.PermissionID) AS ProfileCount,
-                (SELECT COUNT(*) FROM UserPermissionOverride o WHERE o.PermissionID = p.PermissionID) AS OverrideCount
-            FROM Permission p
-            ORDER BY
-                LEFT(p.Code, LEN(p.Code) - CHARINDEX('.', REVERSE(p.Code))),
-                CASE
-                    WHEN p.Code LIKE '%.view'    THEN 1
-                    WHEN p.Code LIKE '%.add'     THEN 2
-                    WHEN p.Code LIKE '%.add.%'   THEN 3
-                    WHEN p.Code LIKE '%.edit%'   THEN 4
-                    WHEN p.Code LIKE '%.delete%' THEN 5
-                    ELSE 6
-                END,
-                p.Code
-        """)
-        perms = [
-            dict(zip([col[0] for col in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify(perms)
-    except Exception as e:
-        current_app.logger.error(f"Error listing permissions: {e}")
-        return jsonify({"error": _("An unexpected error occurred")}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_permission_users(perm_id):
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT Code FROM Permission WHERE PermissionID = ?", (perm_id,))
-        row = cursor.fetchone()
-        if not row:
-            return jsonify({"success": False, "message": _("Permission not found")}), 404
-        perm_code = row[0]
-        cursor.execute(
-            """
-            SELECT
-                u.userID, u.username, u.fullname,
-                ap.Name AS AccessProfileName,
-                CAST(dbo.fnUserHasPermission(u.userID, ?) AS INT) AS HasPermission,
-                upo.Effect AS OverrideEffect,
-                app.Effect AS ProfileEffect
-            FROM Users u
-            LEFT JOIN AccessProfile ap ON ap.AccessID = u.accessID
-            LEFT JOIN UserPermissionOverride upo ON upo.UserID = u.userID AND upo.PermissionID = ?
-            LEFT JOIN AccessProfilePermission app ON app.AccessID = u.accessID AND app.PermissionID = ?
-            ORDER BY u.fullname
-        """,
-            (perm_code, perm_id, perm_id),
-        )
-        users = [
-            dict(zip([col[0] for col in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify({"success": True, "users": users})
-    except Exception as e:
-        current_app.logger.error(f"Error fetching users for permission {perm_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.view.accessprofiles.useroverrides")
-def api_admin_user_all_permissions(user_id):
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT
-                p.PermissionID, p.Code, p.Description,
-                CAST(dbo.fnUserHasPermission(?, p.Code) AS INT) AS IsEffective,
-                upo.Effect AS OverrideEffect,
-                app.Effect AS ProfileEffect
-            FROM Permission p
-            LEFT JOIN Users u ON u.userID = ?
-            LEFT JOIN UserPermissionOverride upo ON upo.UserID = ? AND upo.PermissionID = p.PermissionID
-            LEFT JOIN AccessProfilePermission app ON app.AccessID = u.accessID AND app.PermissionID = p.PermissionID
-            ORDER BY p.Code
-        """,
-            (user_id, user_id, user_id),
-        )
-        perms = [
-            dict(zip([col[0] for col in cursor.description], row, strict=False))
-            for row in cursor.fetchall()
-        ]
-        return jsonify({"success": True, "permissions": perms})
-    except Exception as e:
-        current_app.logger.error(f"Error fetching all permissions for user {user_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.accessprofile")
-def api_admin_permission_add():
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    data = request.get_json()
-    code = (data.get("code") or "").strip()
-    description = (data.get("description") or "").strip()
-    if not code or not description:
-        return jsonify({"success": False, "message": _("Code and description are required")}), 400
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "INSERT INTO Permission (Code, Description) OUTPUT INSERTED.PermissionID VALUES (?, ?)",
-            (code, description),
-        )
-        new_id = cursor.fetchone()[0]
-        conn.commit()
-        return jsonify(
-            {
-                "success": True,
-                "message": _("Permission created successfully"),
-                "permissionId": new_id,
-            }
-        )
-    except Exception as e:
-        current_app.logger.error(f"Error creating permission: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.accessprofile")
-def api_admin_permission_edit(perm_id):
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    data = request.get_json()
-    code = (data.get("code") or "").strip()
-    description = (data.get("description") or "").strip()
-    if not code or not description:
-        return jsonify({"success": False, "message": _("Code and description are required")}), 400
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "UPDATE Permission SET Code=?, Description=? WHERE PermissionID=?",
-            (code, description, perm_id),
-        )
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Permission not found")}), 404
-        conn.commit()
-        return jsonify({"success": True, "message": _("Permission updated successfully")})
-    except Exception as e:
-        current_app.logger.error(f"Error updating permission {perm_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-@require_permission("admin.edit.accessprofile")
-def api_admin_permission_delete(perm_id):
-    if "username" not in session:
-        return jsonify({"error": _("Not authorized")}), 401
-    conn = None
-    cursor = None
-    try:
-        conn = engine_nexora_db.raw_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT COUNT(*) FROM AccessProfilePermission WHERE PermissionID=?", (perm_id,)
-        )
-        profile_refs = cursor.fetchone()[0]
-        cursor.execute(
-            "SELECT COUNT(*) FROM UserPermissionOverride WHERE PermissionID=?", (perm_id,)
-        )
-        override_refs = cursor.fetchone()[0]
-        if profile_refs > 0 or override_refs > 0:
-            return jsonify(
-                {
-                    "success": False,
-                    "message": _(
-                        "Cannot delete: used in %(p)d profile(s) and %(o)d override(s). Remove all assignments first."
-                    )
-                    % {"p": profile_refs, "o": override_refs},
-                }
-            ), 400
-        cursor.execute("DELETE FROM Permission WHERE PermissionID=?", (perm_id,))
-        if cursor.rowcount == 0:
-            return jsonify({"success": False, "message": _("Permission not found")}), 404
-        conn.commit()
-        return jsonify({"success": True, "message": _("Permission deleted successfully")})
-    except Exception as e:
-        current_app.logger.error(f"Error deleting permission {perm_id}: {e}")
-        return jsonify({"success": False, "message": str(e)}), 500
-    finally:
-        if cursor:
-            cursor.close()
-        if conn:
-            conn.close()
-
-
-def register_routes(app):
-    # overview
-    app.add_url_rule("/admin", endpoint="admin_dashboard", view_func=admin_dashboard)
-
-    # organizations
-    app.add_url_rule(
-        "/admin/organizations",
-        endpoint="admin_organizations_view",
-        view_func=admin_organizations_view,
-    )
-    app.add_url_rule(
-        "/admin/organizations/add",
-        endpoint="admin_add_organization",
-        view_func=admin_add_organization,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/organizations/edit/<organizationcode>",
-        endpoint="admin_edit_organization",
-        view_func=admin_edit_organization,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/organizations/delete/<organizationcode>",
-        endpoint="admin_delete_organization",
-        view_func=admin_delete_organization,
-        methods=["DELETE"],
-    )
-    app.add_url_rule(
-        "/api/admin/organizations/list",
-        endpoint="api_admin_organizations_list",
-        view_func=api_admin_organizations_list,
-    )
-    app.add_url_rule(
-        "/admin/organizations/<organizationcode>/branding",
-        endpoint="api_admin_organization_branding_save",
-        view_func=api_admin_organization_branding_save,
-        methods=["POST"],
-    )
-
-    # clients (runtime sources)
-    app.add_url_rule("/admin/clients", endpoint="admin_clients_view", view_func=admin_clients_view)
-    app.add_url_rule(
-        "/admin/clients/add",
-        endpoint="api_admin_clients_add",
-        view_func=api_admin_clients_add,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/clients/edit/<clientcode>",
-        endpoint="api_admin_clients_edit",
-        view_func=api_admin_clients_edit,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/clients/delete/<clientcode>",
-        endpoint="api_admin_clients_delete",
-        view_func=api_admin_clients_delete,
-        methods=["DELETE"],
-    )
-
-    # processes (mapping config)
-    app.add_url_rule(
-        "/admin/processes", endpoint="admin_processes_view", view_func=admin_processes_view
-    )
-    app.add_url_rule(
-        "/api/admin/processes/list",
-        endpoint="api_admin_processes_list",
-        view_func=api_admin_processes_list,
-    )
-    app.add_url_rule(
-        "/admin/processes/sources/add",
-        endpoint="api_admin_process_source_add",
-        view_func=api_admin_process_source_add,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/processes/sources/edit/<clientcode>/<processname>",
-        endpoint="api_admin_process_source_edit",
-        view_func=api_admin_process_source_edit,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/processes/sources/delete/<clientcode>/<processname>",
-        endpoint="api_admin_process_source_delete",
-        view_func=api_admin_process_source_delete,
-        methods=["DELETE"],
-    )
-    app.add_url_rule(
-        "/admin/processes/fields/add",
-        endpoint="api_admin_field_mapping_add",
-        view_func=api_admin_field_mapping_add,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/processes/fields/edit/<clientcode>/<processname>/<fieldkey>",
-        endpoint="api_admin_field_mapping_edit",
-        view_func=api_admin_field_mapping_edit,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/processes/fields/delete/<clientcode>/<processname>/<fieldkey>",
-        endpoint="api_admin_field_mapping_delete",
-        view_func=api_admin_field_mapping_delete,
-        methods=["DELETE"],
-    )
-
-    # status page
-    app.add_url_rule("/admin/status", endpoint="admin_status_view", view_func=admin_status_view)
-
-    # dev server restart
-    app.add_url_rule(
-        "/api/admin/restart",
-        endpoint="api_admin_restart",
-        view_func=api_admin_restart,
-        methods=["POST"],
-    )
-
-    # maintenance banner
-    app.add_url_rule(
-        "/admin/maintenance", endpoint="admin_maintenance_view", view_func=admin_maintenance_view
-    )
-    app.add_url_rule(
-        "/api/admin/maintenance",
-        endpoint="api_admin_maintenance_list",
-        view_func=api_admin_maintenance_list,
-        methods=["GET"],
-    )
-    app.add_url_rule(
-        "/api/admin/maintenance",
-        endpoint="api_admin_maintenance_add",
-        view_func=api_admin_maintenance_add,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/maintenance/<int:banner_id>",
-        endpoint="api_admin_maintenance_edit",
-        view_func=api_admin_maintenance_edit,
-        methods=["PUT"],
-    )
-    app.add_url_rule(
-        "/api/admin/maintenance/<int:banner_id>",
-        endpoint="api_admin_maintenance_delete",
-        view_func=api_admin_maintenance_delete,
-        methods=["DELETE"],
-    )
-
-    # logs
-    app.add_url_rule("/admin/logs", endpoint="admin_logs_view", view_func=admin_logs_view)
-    app.add_url_rule(
-        "/api/admin/logs/search", endpoint="api_admin_logs_search", view_func=api_admin_logs_search
-    )
-    app.add_url_rule(
-        "/api/admin/logs/export.csv",
-        endpoint="api_admin_logs_export",
-        view_func=api_admin_logs_export,
-    )
-
-    # sessions & users
-    app.add_url_rule(
-        "/admin/sessions", endpoint="admin_sessions_view", view_func=admin_sessions_view
-    )
-    app.add_url_rule(
-        "/admin/users/add", endpoint="admin_add_user", view_func=admin_add_user, methods=["POST"]
-    )
-    app.add_url_rule(
-        "/admin/users/edit/<int:user_id>",
-        endpoint="admin_edit_user",
-        view_func=admin_edit_user,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/users/<int:user_id>", endpoint="admin_user_detail", view_func=admin_user_detail
-    )
-    app.add_url_rule(
-        "/api/admin/users/<int:user_id>/activity",
-        endpoint="api_admin_user_activity",
-        view_func=api_admin_user_activity,
-    )
-    app.add_url_rule(
-        "/admin/users/delete/<int:user_id>",
-        endpoint="admin_delete_user",
-        view_func=admin_delete_user,
-        methods=["DELETE"],
-    )
-    app.add_url_rule(
-        "/admin/sessions/<string:session_id>/revoke",
-        endpoint="admin_revoke_session",
-        view_func=admin_revoke_session,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/admin/users/<int:user_id>/revoke_all",
-        endpoint="admin_revoke_all_sessions",
-        view_func=admin_revoke_all_sessions,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/users/list", endpoint="api_admin_users_list", view_func=api_admin_users_list
-    )
-    app.add_url_rule(
-        "/api/admin/recent_logs", endpoint="admin_recent_logs", view_func=admin_recent_logs
-    )
-    app.add_url_rule(
-        "/api/admin/active_sessions",
-        endpoint="admin_active_sessions",
-        view_func=admin_active_sessions,
-    )
-
-    # access control & permissions
-    app.add_url_rule(
-        "/admin/access_control", endpoint="admin_access_control", view_func=admin_access_control
-    )
-    app.add_url_rule(
-        "/admin/permission_matrix",
-        endpoint="admin_permission_matrix",
-        view_func=admin_permission_matrix,
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/<int:permission_id>/holders",
-        endpoint="api_admin_permission_holders",
-        view_func=api_admin_permission_holders,
-    )
-    app.add_url_rule(
-        "/api/admin/users",
-        endpoint="get_users_admin_access_control",
-        view_func=get_users_admin_access_control,
-    )
-    app.add_url_rule(
-        "/api/admin/access_profile/<int:access_id>/details",
-        endpoint="get_profile_details",
-        view_func=get_profile_details,
-        methods=["GET"],
-    )
-    app.add_url_rule(
-        "/api/admin/access_profile/save",
-        endpoint="save_access_profile",
-        view_func=save_access_profile,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/user_overrides/<int:user_id>",
-        endpoint="get_user_overrides",
-        view_func=get_user_overrides,
-        methods=["GET"],
-    )
-    app.add_url_rule(
-        "/api/admin/users/<int:user_id>/effective_permissions",
-        endpoint="api_admin_user_effective_permissions",
-        view_func=api_admin_user_effective_permissions,
-    )
-    app.add_url_rule(
-        "/api/admin/user_overrides/save",
-        endpoint="save_user_overrides",
-        view_func=save_user_overrides,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/list",
-        endpoint="api_admin_permissions_list",
-        view_func=api_admin_permissions_list,
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/<int:perm_id>/users",
-        endpoint="api_admin_permission_users",
-        view_func=api_admin_permission_users,
-    )
-    app.add_url_rule(
-        "/api/admin/users/<int:user_id>/all_permissions",
-        endpoint="api_admin_user_all_permissions",
-        view_func=api_admin_user_all_permissions,
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/add",
-        endpoint="api_admin_permission_add",
-        view_func=api_admin_permission_add,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/edit/<int:perm_id>",
-        endpoint="api_admin_permission_edit",
-        view_func=api_admin_permission_edit,
-        methods=["POST"],
-    )
-    app.add_url_rule(
-        "/api/admin/permissions/delete/<int:perm_id>",
-        endpoint="api_admin_permission_delete",
-        view_func=api_admin_permission_delete,
-        methods=["DELETE"],
-    )
